% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_hmm.R
\name{add_mrk_at_tail_ph_list}
\alias{add_mrk_at_tail_ph_list}
\title{add a single marker at the tail of a linkage phase list}
\usage{
add_mrk_at_tail_ph_list(ph.list.1, ph.list.2, cor.index)
}
\arguments{
\item{void}{internal function}
}
\description{
add a single marker at the tail of a linkage phase list
}
\keyword{internal}
