% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aggregate_matrix}
\alias{aggregate_matrix}
\title{Aggregate matrix cells (lower the resolution by a factor)}
\usage{
aggregate_matrix(M, fact)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Aggregate matrix cells (lower the resolution by a factor)
}
\keyword{internal}
