% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phase_configurations.R
\name{all_phases_with_added_mrk}
\alias{all_phases_with_added_mrk}
\title{Enumerate all phases with an added marker}
\usage{
all_phases_with_added_mrk(prev.info, seq.temp, sh, ploidy, parent, data.name)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Enumerate all phases with an added marker
}
\keyword{internal}
