% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache_twopts.R
\name{cache_counts_twopt}
\alias{cache_counts_twopt}
\title{Frequency of genotypes for two-point estimaton of recombination fractions}
\usage{
cache_counts_twopt(input.seq, get.from.web = FALSE, cache.prev = NULL,
  n.clusters = 1, verbose = TRUE, joint.prob = FALSE)
}
\arguments{
\item{input.seq}{an object of class \code{mappoly.sequence}}

\item{get.from.web}{If \code{TRUE}, access the counts for all
linkage phase configurations in a remote server}

\item{cache.prev}{an object of class \code{cache.info} containing
pre-computed genotype frequencies, obtained with
\code{\link[mappoly]{cache_counts_twopt}} (not used in this version)}

\item{n.clusters}{Number of parallel processes to spawn}

\item{verbose}{If \code{TRUE}, print the linkage phase
confgurations. If \code{get.from.web = TRUE}, nothing is
printed, since all linkage phase configurations will be cached.}

\item{joint.prob}{If \code{FALSE}, returns the frequency of
genotypes for transition probabilities (conditional
probabilities). If \code{TRUE} returns the frequency for joint
probabilities#' @param x an object of one of the classes \code{mappoly.map}}

\item{x}{an object of one of the classes \code{cache.info}}

\item{...}{currently ignored}
}
\value{
An object of class \code{cache.info} which contains a list
    for all pairs of dosages contained in the dataset. The names of
    this list are of the form 'A-B-C-D', where A represents the
    dosage in parent 1, marker k, B represents the dosage in parent
    1, marker k+1, C represents the dosage in parent 2, marker k
    and D represents the dosage in parent 2, marker k+1.  For each
    list, the frequencies were computed for all possible linkage
    phase configurations. The frequencies for each linkage phase
    configuration are distributed in matrices whose names
    represents the number of homologous chromosomes that share
    alleles. The rows on these matrices represents the dosages in k
    and k+1 in an individual in the offspring. See Table 6 of
    Mollinari and Garcia (2017) for an example.
}
\description{
Returns the frequency of each genotype for the two-point reduction
of dimensionality. The frequency is calculated for all pairwise
combination and for all possible linkage phase configuration.
}
\examples{
  \dontrun{
    data(hexafake)
    all.mrk<-make_seq_mappoly(hexafake, 'all')
    ## local computation
    counts<-cache_counts_twopt(all.mrk, n.clusters = 8)
    ## download from web (specially important for high ploidy levels)
    counts.web<-cache_counts_twopt(all.mrk, get.from.web = TRUE)
    }

}
\references{
Mollinari, M., and Garcia, A.  A. F. (2017) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
