% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_genoprob_dist.R
\name{calc_genoprob_dist}
\alias{calc_genoprob_dist}
\title{Compute the conditional QTL probabilities using genotype distribution}
\usage{
calc_genoprob_dist(input.map, dat.dist, phase.config = "best",
  verbose = TRUE)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Compute the conditional QTL probabilities using genotype distribution
}
\keyword{internal}
