% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_map_utils.R
\name{calc_genoprob_haplo}
\alias{calc_genoprob_haplo}
\title{Estimate a genetic map given a sequence of block markers}
\usage{
calc_genoprob_haplo(m, n.mrk, n.ind, haplo, emit = NULL, rf_vec,
  indnames, verbose = TRUE)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Estimate a genetic map given a sequence of block markers
}
\keyword{internal}
