% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{calc_maps_load_and_trim_prin_curve}
\alias{calc_maps_load_and_trim_prin_curve}
\title{Estimates loci position using Principal curves}
\usage{
calc_maps_load_and_trim_prin_curve(lodrf, spar = NULL, n = NULL,
  ndim = 2, weightfn = "lod")
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Estimates loci position using Principal curves
}
\keyword{internal}
