% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{calc_nnfit_loci}
\alias{calc_nnfit_loci}
\title{Get nearest informative marker}
\usage{
calc_nnfit_loci(loci, distmap, lodmap, estmap)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Get nearest informative marker
}
\keyword{internal}
