% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_hmm.R
\name{cat_phase}
\alias{cat_phase}
\title{cat for phase information}
\usage{
cat_phase(
  input.seq,
  input.ph,
  all.ph,
  ct,
  seq.num,
  twopt.phase.number,
  hmm.phase.number
)
}
\description{
cat for phase information
}
\keyword{internal}
