% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_data_sanity}
\alias{check_data_sanity}
\title{Data sanity check}
\usage{
check_data_sanity(x)
}
\arguments{
\item{x}{an object of class \code{mappoly.data}}
}
\value{
if consistent, returns 0. If not consistent, returns a 
        vector with a number of tests, where \code{TRUE} indicates
        a failed test.
}
\description{
Checks the consistency of a dataset
}
\examples{
\donttest{
#### Tetraploid example
fl1 = "https://raw.githubusercontent.com/mmollina/MAPpoly_vignettes/master/data/SolCAP_dosage"
tempfl <- tempfile()
download.file(fl1, destfile = tempfl)
SolCAP.dose <- read_geno(file.in  = tempfl)
check_data_sanity(SolCAP.dose)

#### Hexaploid example
fl2 = "https://raw.githubusercontent.com/mmollina/MAPpoly_vignettes/master/data/hexafake"
tempfl <- tempfile()
download.file(fl2, destfile = tempfl)
hexa.dose <- read_geno(file.in  = tempfl)
check_data_sanity(hexa.dose)
}
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400378}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
