% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_parallel.R
\name{check_gaps}
\alias{check_gaps}
\title{Function to check gaps}
\usage{
check_gaps(map, tpt, max.gap = 0.1, window = 10)
}
\description{
This function was created to facilitate gap checking and weird marker removal.
It is intended to be run in interactive mode.
}
\author{
Gabriel Gesteira, \email{gabrielgesteira@usp.br}
}
\keyword{interval}
