% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_if_rf_is_possible}
\alias{check_if_rf_is_possible}
\title{Check if it is possible to estimate the recombination
fraction between neighbor markers using two-point
estimation}
\usage{
check_if_rf_is_possible(input.seq)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Check if it is possible to estimate the recombination
fraction between neighbor markers using two-point
estimation
}
\keyword{internal}
