% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phase_configurations.R
\name{check_pairwise}
\alias{check_pairwise}
\title{Check if all pairwise combinations of elements of \code{input.seq}
are contained in \code{twopt}}
\usage{
check_pairwise(input.seq, twopt)
}
\arguments{
\item{input.seq}{An object of class \code{mappoly.sequence}}

\item{twopt}{An object of class \code{poly.est.two.pts.pairwise}}
}
\value{
If all pairwise combinations of elements of
    \code{input.seq} are contained in \code{twopt}, the function
    returns 0. Otherwise, returns the missing pairs.
}
\description{
Check if all pairwise combinations of elements of \code{input.seq}
are contained in \code{twopt}
}
\keyword{internal}
