% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_parallel.R
\name{check_phases}
\alias{check_phases}
\title{Check phase conformity}
\usage{
check_phases(map1, map2, verbose = TRUE)
}
\description{
This function is intended to check if all homologues in two given maps
are the same for both parents. The function reduces the sequence
to the common markers only before making the comparison.
}
\author{
Gabriel Gesteira
}
\keyword{internal}
