% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_maps}
\alias{compare_maps}
\title{Compare a list of maps}
\usage{
compare_maps(...)
}
\arguments{
\item{...}{a list of objects of class \code{mappoly.map}}
}
\value{
A data frame where the lines correspond to the maps in the order provided in input list list
}
\description{
Compare lengths, density, maximum gaps and log likelihoods in a list of maps.
In order to make the maps comparable, the function uses the intersection of 
markers among maps.
}
