% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phase_configurations.R
\name{concatenate_new_marker}
\alias{concatenate_new_marker}
\title{Concatenate new marker}
\usage{
concatenate_new_marker(X = NULL, d, sh = NULL, seq.num = NULL, m, mrk = 1)
}
\arguments{
\item{X}{a list of matrices whose columns represent homologous
chromosomes and the rows represent markers}

\item{d}{the dosage of the inserted marker}

\item{sh}{a list of shared alleles between all markers in the sequence}

\item{seq.num}{a vector of integers containing the number of each marker in the raw data file}

\item{m}{the ploidy level}

\item{mrk}{the marker to be inserted}
}
\value{
a unique list of matrices representing linkage phases
}
\description{
Inserts a new marker at the end of the sequence, taking into account
the two-point information
}
\keyword{internal}
