% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_genoprob.R
\name{create_map}
\alias{create_map}
\title{Create a marker with pseudomarkers}
\usage{
create_map(input.map, step = 0, phase.config = "best")
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Create a marker with pseudomarkers
}
\keyword{internal}
