% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_info_par}
\alias{detect_info_par}
\title{Detects which parent is informative}
\usage{
detect_info_par(x)
}
\arguments{
\item{x}{an object of class \code{mappoly.sequence} or \code{mappoly.map}}
}
\description{
Detects which parent is informative
}
