% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_prob_to_class}
\alias{dist_prob_to_class}
\title{Selects the class with high probability and return it instead of
the distribution of probabilities for all classes}
\usage{
dist_prob_to_class(geno, prob.thres = 0.8)
}
\arguments{
\item{geno}{the genotypes contained in the object
\code{'mappoly.data'}}

\item{prob.thres}{probability threshold to accept the genotyupe as
the correct one. Values below this genotype are assumed as
missing data}
}
\value{
a matrix containing the doses of each genotype and
    marker. Missing data are represented by NAs
}
\description{
Selects the class with high probability and return it instead of
the distribution of probabilities for all classes
}
\keyword{internal}
