% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mappoly_vcf.R
\name{.read.vcfR}
\alias{.read.vcfR}
\title{Function read.vcfR adapted from package vcfR}
\usage{
.read.vcfR(
  file,
  limit = 1e+07,
  nrows = -1,
  skip = 0,
  cols = NULL,
  convertNA = TRUE,
  checkFile = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Function read.vcfR adapted from package vcfR
}
\keyword{internal}
