% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_hmm.R
\name{draw_alleles}
\alias{draw_alleles}
\title{draw alleles}
\usage{
draw_alleles(m, x, v, col.cte = 2, y.pos = 0.5)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
draw alleles
}
\keyword{internal}
