% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_hmm.R
\name{draw_homologous}
\alias{draw_homologous}
\title{draw homologous}
\usage{
draw_homologous(m, y.pos = 0.5)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
draw homologous
}
\keyword{internal}
