% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_marker}
\alias{drop_marker}
\title{Remove markers from a map}
\usage{
drop_marker(input.map, mrk)
}
\arguments{
\item{input.map}{an object of class \code{"mappoly.map"}}

\item{mrk}{a vector containing markers to be removed from the input map, identified by their names or positions in the map.}
}
\value{
An object of class 'mappoly.map'
}
\description{
This function creates a new map by removing markers from an existing one.
}
\examples{
\dontrun{
input.map <- maps.hexafake[[1]]
mrk <- c(4, 9, 23)
mrk <- c("M_85", "M_152")
output.map <- drop_marker(input.map, mrk)
plot(output.map, mrk.names = TRUE, cex = .5)
}

}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
