% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_mrk.R
\name{drop_mrk}
\alias{drop_mrk}
\title{Drop markers in a genetic map}
\usage{
drop_mrk(input.map, drop.mrk, reest.map = TRUE, input.mat = NULL,
  tol = 0.01, phase.config = "best", method = c("hmm", "ols"),
  weight = TRUE, verbose = TRUE)
}
\arguments{
\item{input.map}{An object of class \code{mappoly.map}}

\item{drop.mrk}{Markers to be removed from the map}

\item{reest.map}{if \code{TRUE}, reestimate the map using the
method indicated in the argument 'method'}

\item{input.mat}{An object of class \code{mappoly.rf.matrix}}

\item{tol}{tolerance for determining convergence.}

\item{phase.config}{should be a string \code{'best'} or the position of the
configuration to be plotted. If \code{'best'}, plot the configuration
with the highest likelihood.}

\item{method}{indicates whether to use Hidden Markov Models or Ordinary
Least Squares to reestimate the recombination fraction}

\item{weight}{if \code{TRUE}, it uses the LOD scores to perform a weighted
regression when the Ordinary Least Saquares is chosen}

\item{verbose}{If \code{TRUE}, current progress is shown; if
\code{FALSE}, no output is produced.}
}
\description{
Drop markers in a genetic map
}
\examples{
 \dontrun{
    data(hexafake)
    mrk.subset<-make_seq_mappoly(hexafake, 1:100)
    red.mrk<-elim_redundant(mrk.subset)
    unique.mrks<-make_seq_mappoly(red.mrk)
    counts.web<-cache_counts_twopt(unique.mrks, get.from.web = TRUE)
    subset.pairs<-est_pairwise_rf(input.seq = unique.mrks,
                                  count.cache = counts.web,
                                  n.clusters = 16,
                                  verbose=TRUE)
    system.time(
    subset.map <- est_rf_hmm_sequential(input.seq = unique.mrks,
                                        thres.twopt = 5,
                                        thres.hmm = 3,
                                        extend.tail = 50,
                                        tol = 0.1,
                                        tol.final = 10e-3,
                                        twopt = subset.pairs,
                                        verbose = TRUE))
     plot(subset.map)
     #removing markers 'M10', 'M15' and 'M24'
     new.map<-drop_mrk(input.map = subset.map,
                       drop.mrk = c(10,15,24),
                       reest.map = TRUE,
                       tol = 10e-3,
                       method = "hmm")
     plot(new.map)
 }

}
\references{
Mollinari, M., and Garcia, A.  A. F. (2018) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
