% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{edit_order}
\alias{edit_order}
\title{Edit sequence ordered by reference genome positions 
comparing to another set order}
\usage{
edit_order(input.seq, input.data)
}
\arguments{
\item{input.seq}{object of class mappoly.sequence with alternative order (not genomic order)}

\item{input.data}{object of class mappoly.data}
}
\description{
Edit sequence ordered by reference genome positions 
comparing to another set order
}
\author{
Cristiane Taniguti, \email{chtaniguti@tamu.edu}
}
