% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phase_configurations.R
\name{elim_conf_using_two_pts}
\alias{elim_conf_using_two_pts}
\title{Eliminate configurations using two-point information}
\usage{
elim_conf_using_two_pts(input.seq, twopt, thres)
}
\arguments{
\item{input.seq}{an object of class \code{mappoly.sequence}.}

\item{twopt}{an object of class \code{poly.est.two.pts.pairwise}}

\item{thres}{threshold from which the linkage phases can be
discarded (if abs(ph_LOD) > thres)}
}
\value{
a unique list of matrices representing linkage phases
}
\description{
Drops unlikely configuration phases given the two-point information
and a LOD threshold
}
\keyword{internal}
