% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phase_configurations.R
\name{elim_equiv}
\alias{elim_equiv}
\title{Eliminates equivalent linkage phase configurations}
\usage{
elim_equiv(Z)
}
\arguments{
\item{Z}{a list of matrices whose columns represent homologous
chromosomes and the rows represent markers}
}
\value{
a unique list of matrices
}
\description{
Drop equivalent linkage phase configurations, i.e. the ones which
have permuted homologous chromosomes
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
