% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{elim_redundant_in_datasets}
\alias{elim_redundant_in_datasets}
\title{Eliminate redundant markers in datasets}
\usage{
elim_redundant_in_datasets(dat, is.vcf = FALSE)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Eliminate redundant markers in datasets
}
\keyword{internal}
