% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_map_with_genotyping_error.R
\name{est_full_hmm_with_global_error}
\alias{est_full_hmm_with_global_error}
\title{Reestimate gentic map given a global genotyping error}
\usage{
est_full_hmm_with_global_error(input.map, error = NULL, tol = 0.001)
}
\arguments{
\item{input.map}{an object of class \code{mappoly.map}.}

\item{error}{global error rate}

\item{tol}{the desired accuracy.}
}
\value{
An object of class 'mappoly.map'
}
\description{
This function considers a global error when reestimating
a genetic map using hdden Markov models
}
\examples{
  \dontrun{
    hexa_file<-system.file('extdata', 'hexa_fake', package = 'mappoly2')
    hexa_dat<-read_geno(file_in = hexa_file)
    all_mrk<-make_seq_mappoly(hexa_dat, 'all')
    counts_all_mrk_from_web<-cache_counts_twopt(all_mrk,
                                                get.from.web=TRUE)
    all_pairs<-est_pairwise_rf(all_mrk, counts_all_mrk_from_web,
                               n.clusters=2)

    seq1.10<-make_seq_mappoly(hexa_dat, 1:10)

    l_seq1_3.0 <- ls_linkage_phases(input.seq = seq1.10, thres = 1.3,
                                    twopt = all_pairs)
    l_seq1_3.0
    plot(l_seq1_3.0)
    res<-est_rf_hmm(input.seq=seq1.10, input.ph=l_seq1_3.0,
                    verbose=TRUE, tol=10e-3)
    res
    plot(res)

    res_final<-est_full_hmm_with_global_error(res, error=0.01, tol=10e-4)
    plot(res_final)
    }

}
\references{
Mollinari, M., and Garcia, A.  A. F. (2018) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
