% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/reestimate_map_with_prior_genotypic_distribution.R
\name{est_full_hmm_with_prior_dist}
\alias{est_full_hmm_with_prior_dist}
\title{Reestimate gentic map using dosage prior distribution}
\usage{
est_full_hmm_with_prior_dist(input.map, dat.dist = NULL,
  phase.config = "best", tol = 0.001, verbose = TRUE)
}
\arguments{
\item{input.map}{an object of class \code{mappoly.map}.}

\item{dat.dist}{an object of class \code{mappoly.data} containing 
the probability distribution of the genotypes. If 
\code{NULL} (default), it uses the original data 
set used to construct \code{input.map}.}

\item{phase.config}{which phase configuration should be used
"best" will choose the one with highest likelihood}

\item{tol}{the desired accuracy.}

\item{verbose}{if \code{TRUE}, current progress is shown; if
\code{FALSE}, no output is produced.}
}
\value{
An object of class 'mappoly.map'
}
\description{
This function considers dosage prior distribution when reestimating
a genetic map using hdden Markov models
}
\examples{
  \dontrun{
  
  solcap.p<-vector("list", 12)
  names(solcap.p)<-names(solcap.dose.map)
  for(i in 1:12)
    solcap.p[[i]] <- est_full_hmm_with_prior_dist(solcap.dose.map[[i]], verbose = FALSE)
 w<-NULL
 for(i in 1:12)
   w<-c(w, c(solcap.dose.map[i], 
             solcap.p[i]))
             
 names(w) <- apply(expand.grid(c("dose", "prior"), paste0("LG_", 1:12), 
                             stringsAsFactors = FALSE)[,2:1], 1, paste, 
                 collapse = "_")
                 
 op <- par(cex.axis = .7)
 plot_map_list(w, horiz = FALSE, col = rep(gg_color_hue(2), 12))
 par(op)
 legend("bottomright", legend = c("Dosage based", "Prior"), pch=15, col = rep(gg_color_hue(2)))
  
}

}
\references{
Mollinari, M., and Garcia, A.  A. F. (2018) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_. \url{https://doi.org/10.1101/415232}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
