% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_parallel.R
\name{est_map_parallel}
\alias{est_map_parallel}
\title{Function to build a map in parallel with submaps}
\usage{
est_map_parallel(
  data,
  markers,
  partial_tpt,
  n.batches = 4,
  n.cores = 4,
  platform = "auto",
  start.set = 5,
  thres.twopt = 10,
  thres.hmm = 10,
  submap.size.diff = 3,
  thres.twopt2 = 3,
  thres.hmm2 = 10,
  phase.n.lim = 20,
  tol = 0.001
)
}
\arguments{
\item{void}{}
}
\description{
This function estimates the map for a sequence using batches (submaps)
and parallel computing.
}
\author{
Gabriel Gesteira, \email{gabrielgesteira@usp.br}
}
\keyword{internal}
