% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_rf.R
\name{est_pairwise_rf}
\alias{est_pairwise_rf}
\alias{print.poly.est.two.pts.pairwise}
\alias{plot.poly.est.two.pts.pairwise}
\title{Pairwise two-point analysis}
\usage{
est_pairwise_rf(input.seq, count.cache, n.clusters = 1,
  tol = .Machine$double.eps^0.25, mrk.pairs = NULL,
  batch.size = NULL, verbose = TRUE)

\method{print}{poly.est.two.pts.pairwise}(x, ...)

\method{plot}{poly.est.two.pts.pairwise}(x, first.mrk, second.mrk, ...)
}
\arguments{
\item{input.seq}{an object of class \code{mappoly.sequence}.}

\item{count.cache}{an object of class \code{cache.info} containing
pre-computed genotype frequencies, obtained with
\code{\link[mappoly]{cache_counts_twopt}}.}

\item{n.clusters}{Number of parallel processes to spawn}

\item{tol}{the desired accuracy. See \code{optimize()} for details.}

\item{mrk.pairs}{a matrix of dimensions $2 x n$, containing $n$
pairs of markers to be analyzed. If \code{NULL}, all pairs are
considered}

\item{batch.size}{The number of pairs that shold be analyzed in parallel.
If \code{NULL}, all pairs are analyized.}

\item{verbose}{If \code{TRUE}, current progress is shown; if
\code{FALSE}, no output is produced.}

\item{x}{an object of one of the classes \code{poly.est.two.pts.pairwise}}

\item{...}{currently ignored}

\item{first.mrk}{index indicating the position of the first marker}

\item{second.mrk}{index indicating the position of the second marker}
}
\value{
An object of class \code{poly.est.two.pts.dosage.pairwise} which
    is a list  containing the following components:
    \item{data.name}{name of the object of class
        \code{mappoly.data} with the raw data.}
    \item{n.mrk}{number of markers in the sequence}
    \item{seq.num}{a \code{vector} containing the (ordered) indices
        of markers inthe sequence, according to the input file.}
    \item{pairwise}{which is a list of size}
    \code{choose(length(input.seq$seq.num), 2)} containing a matrix
    with the linkage phase configuration i.e. the number of
    homologous that share alelels, the LOD Score relative to eacho
    oneof these configurations, the recombination fraction
    estimated for each confoguration and their associated LOD
    score.
}
\description{
Performs the two-point pairwise analysis.
}
\examples{
  \dontrun{
    data(hexafake)
    all.mrk<-make_seq_mappoly(hexafake, 'all')
    red.mrk<-elim_redundant(all.mrk)
    unique.mrks<-make_seq_mappoly(red.mrk)
    counts.web<-cache_counts_twopt(unique.mrks, get.from.web = TRUE)
    all.pairs<-est_pairwise_rf(input.seq = unique.mrks,
                               count.cache = counts.web,
                               n.clusters = 16,
                               verbose=TRUE)
   all.pairs
   plot(all.pairs, 90, 91)
   }
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2018) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_. \url{https://doi.org/10.1101/415232}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
