% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_hmm.R
\name{est_rf_hmm}
\alias{est_rf_hmm}
\title{Multipoint analysis using Hidden Markov Models in autopolyploids}
\usage{
est_rf_hmm(input.seq, input.ph = NULL, thres = 0.5, twopt = NULL,
  verbose = FALSE, rf.lim = 0.5, tol = 1e-04, est.given.0.rf = FALSE,
  reestimate.single.ph.configuration = TRUE)
}
\arguments{
\item{input.seq}{an object of class \code{mappoly.sequence}.}

\item{input.ph}{an object of class \code{two.pts.linkage.phases}.}

\item{thres}{the threshold used to determine if the linkage phases
compared via two-point analysis should be considered}

\item{twopt}{an object of class \code{poly.est.two.pts.pairwise}
containing the two-point information}

\item{verbose}{if \code{TRUE}, current progress is shown; if
\code{FALSE}, no output is produced.}

\item{rf.lim}{limit of recombination fraction during convergence.
This is specially useful for marker blocks.}

\item{tol}{the desired accuracy.}

\item{est.given.0.rf}{logical. If TRUE returns a map forcing all
recombination fractions equals to 0 (1e-5)}

\item{reestimate.single.ph.configuration}{logical. If \code{TRUE}
returns a map without reestimating the map parameters for cases
where there are only one possible linkage phase configuration. This argument
is intended to be used in a sequential map contruction.}

\item{x}{an object of one of the classes \code{mappoly.map}}

\item{config}{should be a string \code{'best'} or the position of the
configuration to be plotted. If \code{'best'}, plot the configuration
with the highest likelihood.}

\item{...}{currently ignored}
}
\value{
An object of class 'mappoly.map'
}
\description{
Performs the multipoint analysis proposed by \cite{Mollinari and
 Garcia et al.  (2017)} in a sequence of markers
}
\details{
This function first enumerates a set of linkage phase configurations
 based on two-point recombination fraction information using a threshold
 provided by the user (argument \code{'thresh'}). After that, for each
 one of the configurations, it reconstructs the genetic map for using the
 HMM approach. As result, it returns the multipoint likelihood for each
 configuration in form of LOD Score comparing each configuration to the most
 likely one. It is recommended to use a small number of markers (e.g. 50 markers
 for hexaploids) since the possible linkage phase combinations bounded only
 by the two-point information can be huge. Also, it can be quite sensible
 to small changes in \code{'thresh'}. Thus it is strongly recommended use a small
 \code{'thresh'} value in the first run (0.1 - 0.5) and raises it to find a
 reasonable number of linkage phase configurations to be evaluated by the
 HMM approach. For higher number of markers, please see  \code{'est_rf_hmm_sequential'}
 and \code{'est_map_haplotype'}
}
\examples{
 \dontrun{
    data(hexafake)
    mrk.subset<-make_seq_mappoly(hexafake, 1:50)
    red.mrk<-elim_redundant(mrk.subset)
    unique.mrks<-make_seq_mappoly(red.mrk)
    counts.web<-cache_counts_twopt(unique.mrks, get.from.web = TRUE)
    subset.pairs<-est_pairwise_rf(input.seq = unique.mrks,
                                  count.cache = counts.web,
                                  n.clusters = 1,
                                  verbose=TRUE)

    ## Estimating subset map with a low tolerance for the E.M. procedure
    subset.map <- est_rf_hmm(input.seq = unique.mrks,
                             thres = 2,
                             twopt = subset.pairs,
                             verbose = TRUE,
                             rf.lim = 0.5,
                             tol = 0.1,
                             est.given.0.rf = FALSE)

    ## Re-estimating the map with the most likely configuration
    subset.map1 <- est_rf_hmm_single(input.seq = unique.mrks,
                                    input.ph.single = subset.map$maps[[1]]$seq.ph,
                                    tol = 10e-4,
                                    verbose = TRUE)

    subset.map$maps[[1]]$seq.ph <- subset.map1$seq.ph

     ## Retrieving simulated linkage phase
     sim.ph.P.file <- system.file('doc', 'phase_sim_hexa_P.csv', package = 'mappoly')
     ph.P <- read.csv2(sim.ph.P.file)
     ph.P <- ph_matrix_to_list(ph.P[1:50,-1])
     sim.ph.Q.file <- system.file('doc', 'phase_sim_hexa_Q.csv', package = 'mappoly')
     ph.Q <- read.csv2(sim.ph.Q.file)
     ph.Q <- ph_matrix_to_list(ph.Q[1:50,-1])

     ## Estimated linkage phase
     ph.P.est <- subset.map$maps[[1]]$seq.ph$P
     ph.Q.est <- subset.map$maps[[1]]$seq.ph$Q

     compare_haplotypes(m = 6, h1 = ph.P[names(ph.P.est)], h2 = ph.P.est)
     compare_haplotypes(m = 6, h1 = ph.Q[names(ph.Q.est)], h2 = ph.Q.est)

     try(dev.off(), silent = TRUE)
     plot(subset.map, 1)

   }

}
\references{
Mollinari, M., and Garcia, A.  A. F. (2017) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
