% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_map_hmm.R
\name{est_rf_hmm_single}
\alias{est_rf_hmm_single}
\title{Multipoint analysis using Hidden Markov Models (single phase)}
\usage{
est_rf_hmm_single(input.seq, input.ph.single, rf.temp = NULL, tol,
  verbose = FALSE, ret.map.no.rf.estimation = FALSE,
  high.prec = TRUE)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Multipoint analysis using Hidden Markov Models (single phase)
}
\keyword{internal}
