% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_paprent_single_map_hmm.R
\name{est_rf_hmm_single_one_parent}
\alias{est_rf_hmm_single_one_parent}
\title{Multipoint analysis using Hidden Markov Models (single phase)}
\usage{
est_rf_hmm_single_one_parent(
  input.seq,
  input.ph.single,
  info.parent = 1,
  uninfo.parent = 2,
  rf.vec = NULL,
  global.err = 0,
  tol = 0.001,
  verbose = FALSE,
  ret.map.no.rf.estimation = FALSE
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Multipoint analysis using Hidden Markov Models (single phase)
}
\examples{
  \donttest{
    seq.all.mrk <- make_seq_mappoly(hexafake, 1:20)
 }
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
