% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_map_hmm.R
\name{est_rf_hmm_single_phase}
\alias{est_rf_hmm_single_phase}
\title{Multipoint analysis using Hidden Markov Models (single phase)}
\usage{
est_rf_hmm_single_phase(
  input.seq,
  input.ph.single,
  rf.temp = NULL,
  tol,
  verbose = FALSE,
  ret.map.no.rf.estimation = FALSE,
  high.prec = TRUE,
  max.rf.to.break.EM = 0.5
)
}
\description{
Multipoint analysis using Hidden Markov Models (single phase)
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
