% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_paprent_single_phase_hmm.R
\name{est_rf_hmm_single_phase_single_parent}
\alias{est_rf_hmm_single_phase_single_parent}
\title{Multilocus analysis using Hidden Markov Models (single parent, single phase)}
\usage{
est_rf_hmm_single_phase_single_parent(
  input.seq,
  input.ph.single,
  info.parent = 1,
  uninfo.parent = 2,
  rf.vec = NULL,
  global.err = 0,
  tol = 0.001,
  verbose = FALSE,
  ret.map.no.rf.estimation = FALSE
)
}
\description{
Multilocus analysis using Hidden Markov Models (single parent, single phase)
}
\keyword{internal}
