% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_data_to_polymapR}
\alias{export_data_to_polymapR}
\title{Export data to \code{polymapR}}
\usage{
export_data_to_polymapR(data.in)
}
\arguments{
\item{data.in}{an object of class \code{mappoly.data}}
}
\value{
a dosage \code{matrix}
}
\description{
Export data to \code{polymapR}
}
\examples{
\donttest{
require(polymapR)
dat<-export_data_to_polymapR(hexafake)
F1checked <- checkF1(dosage_matrix = dat, 
                     parent1 = "P1",
                     parent2 = "P2",
                     F1 = colnames(dat)[-c(1:2)],
                     polysomic = TRUE, 
                     disomic = FALSE, 
                     mixed = FALSE, 
                     ploidy = 6)
 head(F1checked$checked_F1)
}  
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
