% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_map_list.R
\name{export_map_list}
\alias{export_map_list}
\title{Export a genetic map to a CSV file}
\usage{
export_map_list(map.list, file = "map_output.csv")
}
\arguments{
\item{map.list}{A list of objects or a single object of class \code{mappoly.map}}

\item{file}{either a character string naming a file or a connection open for writing. 
"" indicates output to the console.}
}
\description{
Function to export genetic linkage map(s) generated by \code{MAPpoly}.
The map(s) should be passed as a single object or a list of objects of class \code{mappoly.map}.
}
\examples{
 export_map_list(solcap.err.map[[1]], file = "")

}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400378}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
