% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_list.R
\name{extract_map}
\alias{extract_map}
\title{extract the distance from an object of class 'mappoly.map'}
\usage{
extract_map(input.map, phase.config = "best")
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
extract the distance from an object of class 'mappoly.map'
}
\keyword{internal}
