% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_aneuploid}
\alias{filter_aneuploid}
\title{Filter aneuploid chromosomes from progeny individuals}
\usage{
filter_aneuploid(input.data, aneuploid.info, ploidy)
}
\arguments{
\item{input.data}{name of input object (class \code{mappoly.data})}

\item{aneuploid.info}{data.frame with ploidy information by chromosome (columns) for each individual in progeny (rows)}

\item{ploidy}{main ploidy}
}
\description{
Filter aneuploid chromosomes from progeny individuals
}
\author{
Cristiane Taniguti, \email{chtaniguti@tamu.edu}
}
