% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_individuals}
\alias{filter_individuals}
\title{Filter out individuals}
\usage{
filter_individuals(
  input.data,
  ind.to.remove = NULL,
  inter = TRUE,
  type = c("Gmat", "PCA"),
  verbose = TRUE
)
}
\arguments{
\item{input.data}{name of input object (class \code{mappoly.data})}

\item{ind.to.remove}{individuals to be removed. If \code{NULL} it opens
an interactive graphic to proceed with the individual
selection}

\item{inter}{if \code{TRUE}, expects user-input to proceed with filtering}

\item{type}{A character string specifying the procedure to be used for 
            detecting outlier offspring. Options include "Gmat", which 
            utilizes the genomic kinship matrix, and "PCA", which employs 
            principal component analysis on the dosage matrix.

  coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.}

\item{verbose}{if \code{TRUE} (default), shows the filtered out individuals}
}
\description{
This function removes individuals from the data set. Individuals can be
user-defined or can be accessed via interactive kinship analysis.
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
