% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_marker_blocks.R
\name{filter_marker_blocks}
\alias{filter_marker_blocks}
\title{Filter bins given a recombination fraction threshold}
\usage{
filter_marker_blocks(input.obj, rf.thres)
}
\arguments{
\item{input.obj}{an object of class \code{mappoly.bins}}

\item{rf.thres}{recombination fraction threshold used to eliminate
bins}
}
\description{
Given a object of class \code{mappoly.bins}, returns a new object
of class \code{mappoly.bins} eliminating the bins whose
recombination fractions exceed a given threshold
}
