% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_missing}
\alias{filter_missing}
\title{Filter missing genotypes}
\usage{
filter_missing(input.data, filter.thres = 0.2, inter = TRUE)
}
\arguments{
\item{input.data}{an object of class \code{"mappoly.data"}}

\item{filter.thres}{maximum percentage of missing data}

\item{inter}{if \code{TRUE}, plots markes vs. frequency of genotyped}
}
\description{
Filter missing genotypes
}
\keyword{internal}
