% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_missing_ind}
\alias{filter_missing_ind}
\title{Filter individuals based on missing genotypes}
\usage{
filter_missing_ind(input.data, filter.thres = 0.2, inter = TRUE)
}
\arguments{
\item{input.data}{an object of class \code{"mappoly.data"}}

\item{filter.thres}{maximum percentage of missing data}

\item{inter}{if \code{TRUE}, plots markers vs. frequency of genotyped individuals}
}
\description{
Filter individuals based on missing genotypes
}
\keyword{internal}
