% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_segregation}
\alias{filter_segregation}
\title{marker filter based on chi-square test}
\usage{
filter_segregation(input.data, chisq.pval.thres = 1e-04, inter = TRUE)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
marker filter based on chi-square test
}
\keyword{internal}
