% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_hmm.R
\name{format_name}
\alias{format_name}
\title{format marker names to print}
\usage{
format_name(y, big.name)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
format marker names to print
}
\keyword{internal}
