% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_rf.R
\name{format_rf}
\alias{format_rf}
\title{Format results from pairwise two-point estimation in C++}
\usage{
format_rf(res)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Format results from pairwise two-point estimation in C++
}
\keyword{internal}
