% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_parallel.R
\name{func.merge}
\alias{func.merge}
\title{Wrapper for function merge_maps}
\usage{
func.merge(x, tpt, thres.twopt2, thres.hmm2, tol)
}
\arguments{
\item{void}{}
}
\description{
Wrapper for function merge_maps
}
\keyword{internal}
