% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_parallel.R
\name{func.sequential}
\alias{func.sequential}
\title{Wrapper for function est_rf_hmm_sequential}
\usage{
func.sequential(x, start.set, thres.twopt, thres.hmm, submap.size.diff,
  phase.n.lim, tol)
}
\arguments{
\item{void}{}
}
\description{
Wrapper for function est_rf_hmm_sequential
}
\keyword{internal}
