% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_map_utils.R
\name{generate_all_link_phases_elim_equivalent_haplo}
\alias{generate_all_link_phases_elim_equivalent_haplo}
\title{Generates all possible linkage phases between two markers blocks
eliminating evivalent configurations, i.e., configurations with the
same likelihood}
\usage{
generate_all_link_phases_elim_equivalent_haplo(block1, block2, M, m,
  max.inc = NULL)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Generates all possible linkage phases between two markers blocks
eliminating evivalent configurations, i.e., configurations with the
same likelihood
}
\keyword{internal}
