% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mappoly_vcf.R
\name{genotype_SM}
\alias{genotype_SM}
\title{Function written to perform SNP genotyping using SuperMASSA}
\usage{
genotype_SM(data, ploidy, parent.1, parent.2)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Function written to perform SNP genotyping using SuperMASSA
}
\keyword{internal}
