% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reestimate_map_with_global_genotypic_error.R
\name{genotyping_global_error}
\alias{genotyping_global_error}
\title{Introduces genotyping error into a full informative vector}
\usage{
genotyping_global_error(x, m, restricted = TRUE, error = 0.01, th.prob = 0.95)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
If \code{restricted = TRUE}, it restricts the prior to the 
possible classes under mendelian non double-reduced segregation 
given dosage of the parents
}
\keyword{internal}
