% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts_from_web.R
\name{get_cache_two_pts_from_web}
\alias{get_cache_two_pts_from_web}
\title{Access a remote server to get Counts for recombinant classes}
\usage{
get_cache_two_pts_from_web(
  m,
  url.address = NULL,
  joint.prob = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Access a remote server to get Counts for recombinant classes
}
\keyword{internal}
