% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_counts.R
\name{get_counts_two_parents}
\alias{get_counts_two_parents}
\title{Counts for recombinant classes}
\usage{
get_counts_two_parents(
  x = c(2, 2),
  ploidy,
  p.k,
  p.k1,
  q.k,
  q.k1,
  verbose = FALSE,
  joint.prob = FALSE
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Counts for recombinant classes
}
\keyword{internal}
