% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{get_dist_loci}
\alias{get_dist_loci}
\title{get_dist_loci}
\usage{
get_dist_loci(loci, estmap, realmap)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
get_dist_loci
}
\keyword{internal}
