% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_hmm.R
\name{get_full_info_tail}
\alias{get_full_info_tail}
\title{Get the tail of a marker sequence up to the point where the markers
provide no additional information.}
\usage{
get_full_info_tail(input.obj, extend = NULL)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Get the tail of a marker sequence up to the point where the markers
provide no additional information.
}
\keyword{internal}
