% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_genomic_order}
\alias{get_genomic_order}
\title{get genomic order of the markers}
\usage{
get_genomic_order(input.seq)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
get genomic order of the markers
}
\keyword{internal}
