% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_LOD}
\alias{get_LOD}
\title{Extract the LOD Scores in a \code{'mappoly.map'} object}
\usage{
get_LOD(x, sorted = TRUE)
}
\arguments{
\item{x}{an object of class \code{mappoly.map}}

\item{sorted}{logical. if \code{TRUE}, the LOD Scores are displayed
in a decreasing order}
}
\value{
a numeric vector containing the LOD Scores
}
\description{
Extract the LOD Scores in a \code{'mappoly.map'} object
}
\keyword{internal}
