% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{get_nearest_informative}
\alias{get_nearest_informative}
\title{Get nearest informative marker}
\usage{
get_nearest_informative(loci, lodmap)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Get nearest informative marker
}
\keyword{internal}
