% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ols_map}
\alias{get_ols_map}
\title{Get weighted ordinary least squared map give a sequence and rf matrix}
\usage{
get_ols_map(input.seq, input.map, weight = TRUE)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Get weighted ordinary least squared map give a sequence and rf matrix
}
\keyword{internal}
