% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_hmm.R
\name{get_ph_list_subset}
\alias{get_ph_list_subset}
\title{subset of a linkage phase list}
\usage{
get_ph_list_subset(ph.list, seq.num, conf)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
subset of a linkage phase list
}
\keyword{internal}
