% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phase_configurations.R
\name{get_rf_from_list}
\alias{get_rf_from_list}
\title{Get the recombination fraction for a sequence of markers given an
object of class \code{poly.est.two.pts.pairwise} and a list
containing the linkage phase configuration. This list can be found
in any object of class \code{two.pts.linkage.phases}, in
x$config.to.test$'Conf-i', where x is the object of class
\code{two.pts.linkage.phases} and i is one of the possible
configurations.}
\usage{
get_rf_from_list(twopt, ph.list)
}
\arguments{
\item{twopt}{an object of class \code{poly.est.two.pts.pairwise}}

\item{ph.list}{a list containing the linkage phase configuration. This
list can be found in any object of class
\code{two.pts.linkage.phases}, in x$config.to.test$'Conf-i',
where x is the object of class \code{two.pts.linkage.phases}
and i is one of the possible configurations.}
}
\value{
a vector with the recombination fraction between markers
    present in ph.list, for that specific order.
}
\description{
Get the recombination fraction for a sequence of markers given an
object of class \code{poly.est.two.pts.pairwise} and a list
containing the linkage phase configuration. This list can be found
in any object of class \code{two.pts.linkage.phases}, in
x$config.to.test$'Conf-i', where x is the object of class
\code{two.pts.linkage.phases} and i is one of the possible
configurations.
}
\keyword{internal}
