% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rf_from_mat}
\alias{get_rf_from_mat}
\title{Get recombination fraction from a matrix}
\usage{
get_rf_from_mat(M)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Get recombination fraction from a matrix
}
\keyword{internal}
