% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_states_and_emission_one_parent}
\alias{get_states_and_emission_one_parent}
\title{Get states and emission in one informative parent}
\usage{
get_states_and_emission_one_parent(ploidy, ph, global.err, D)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Get states and emission in one informative parent
}
\keyword{internal}
