% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_states_and_emission_single_parent}
\alias{get_states_and_emission_single_parent}
\title{Get states and emission in one informative parent}
\usage{
get_states_and_emission_single_parent(ploidy, ph, global.err, D, dose.notinf.P)
}
\description{
Get states and emission in one informative parent
}
\keyword{internal}
