% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_submap.R
\name{get_submap}
\alias{get_submap}
\title{Extract sub-map from map}
\usage{
get_submap(input.map, mrk.pos, phase.config = "best", reestimate.rf = TRUE,
  reestimate.phase = FALSE, thres.twopt = 5, thres.hmm = 3,
  extend.tail = 50, count.cache = NULL, tol = 0.1, tol.final = 0.001,
  use.high.precision = TRUE, verbose = TRUE)
}
\arguments{
\item{input.map}{An object of class \code{mappoly.map}}

\item{mrk.pos}{positions of the markers that should be considered in the new map. This can be in any order.}

\item{phase.config}{which phase configuration should be used
"best" will choose the one with highest likelihood}

\item{reestimate.rf}{logical. If \code{TRUE} (default) the recombination fractions between markers are reestimated.}

\item{reestimate.phase}{logical. If \code{TRUE}, the linkage phase configuration is reestimated. The default is \code{FALSE}}

\item{thres.twopt}{the threshold used to determine if the linkage
phases compared via two-point analysis should be considered}

\item{thres.hmm}{the threshold used to determine if the linkage
phases compared via hmm analysis should be considered}

\item{extend.tail}{the length of the tail of the chain that should
be used to calculate the likelihood of the linkage phases. If
\code{info.tail = TRUE}, the function uses at least \code{extend.tail}
as the lengthof the tail.}

\item{count.cache}{an object of class \code{cache.info} containing
pre-computed genotype frequencies, obtained with
\code{\link[mappoly]{cache_counts_twopt}}.}

\item{tol}{the desired accuracy during the sequential phase.}

\item{tol.final}{the desired accuracy for the final map.}

\item{use.high.precision}{If \code{TRUE}, uses a high precision double
variable in C++ routine. This is advised for maps with a large number
of markers.}

\item{verbose}{If \code{TRUE}, current progress is shown; if
\code{FALSE}, no output is produced.}
}
\description{
Given a pre-constructed map, it extracts a sub-map for a sequence of marker positions.
}
\examples{
 \dontrun{
    data(hexafake)
    ##100 first markers as an example
    s1<-make_seq_mappoly(hexafake, 1:100)
    red.mrk<-elim_redundant(s1)
    unique.mrks<-make_seq_mappoly(red.mrk)
    counts.web<-cache_counts_twopt(unique.mrks, get.from.web = TRUE)
    unique.pairs<-est_pairwise_rf(input.seq = unique.mrks,
                                  count.cache = counts.web,
                                  n.clusters = 1,
                                  verbose=TRUE)
    map <- est_rf_hmm_sequential(input.seq = unique.mrks,
                                        thres.twopt = 5,
                                        thres.hmm = 3,
                                        extend.tail = 50,
                                        tol = 0.1,
                                        tol.final = 10e-2,
                                        twopt = subset.pairs,
                                        verbose = TRUE)
  #MDS whithout reestimating phase
  m1<-rf_list_to_matrix(input.twopt = unique.pairs)
  mds.ord <- mds_mappoly(input.mat = m1, p = NULL, n = NULL, ndim = 2)
  mds.seq.ord <- make_seq_mappoly(mds.ord)
  mds.map<-get_submap(input.map = map, mrk.pos = order(mds.seq.ord$seq.num),
                      verbose = TRUE, tol.final = 10e-2)
  plot(map)
  plot(mds.map)
  map$maps[[1]]$loglike
  mds.map$maps[[1]]$loglike
  plot(map$maps[[1]]$seq.num, mds.map$maps[[1]]$seq.num)
  }

}
\references{
Mollinari, M., and Garcia, A.  A. F. (2018) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_. \url{https://doi.org/10.1101/415232}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
