% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tab_mrks}
\alias{get_tab_mrks}
\title{Get table of dosage combinations}
\usage{
get_tab_mrks(x)
}
\arguments{
\item{x}{an object of class \code{mappoly.map}}
}
\description{
Internal function
}
\author{
Gabriel Gesteira, \email{gabrielgesteira@usp.br}
}
