% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{Color pallet ggplot-like}
\usage{
gg_color_hue(n)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Color pallet ggplot-like
}
\keyword{internal}
