% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_map_utils.R
\name{haplotype_index}
\alias{haplotype_index}
\title{Generates a list where each element represents one individual. Each
element of the list contains the states that should be visited
given a marker block}
\usage{
haplotype_index(ph, conf = 1)
}
\arguments{
\item{void}{intern function to be documented}
}
\description{
Generates a list where each element represents one individual. Each
element of the list contains the states that should be visited
given a marker block
}
\keyword{internal}
