% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_rf.R
\name{ll_twopt}
\alias{ll_twopt}
\title{Compute the log-likelihood of a set of pairwise markers given a recombination fraction}
\usage{
ll_twopt(rec.frac, input.seq)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Compute the log-likelihood of a set of pairwise markers given a recombination fraction
}
