% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglike_hmm.R
\name{loglike_hmm}
\alias{loglike_hmm}
\title{Multipoint log-likelihood computation}
\usage{
loglike_hmm(input.map, verbose = FALSE)
}
\arguments{
\item{input.map}{An object of class \code{mappoly.map}}

\item{verbose}{If \code{TRUE}, map information is shown; if
\code{FALSE}(default), no output is produced}
}
\description{
Update the multipoint log-likelihood of a given map 
using the method proposed by \cite{Mollinari and Garcia (2019)}.
}
\examples{
 \dontrun{
  hexa.map1<-reest_rf(solcap.dose.map[[1]], verbose = FALSE, tol = 10e-4)
  hexa.map2<-loglike_hmm(hexa.map1)
  hexa.map1;hexa.map2
 }
 
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400378}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
