% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phase_configurations.R
\name{ls_linkage_phases}
\alias{ls_linkage_phases}
\alias{print.two.pts.linkage.phases}
\alias{plot.two.pts.linkage.phases}
\title{List of linkage phases}
\usage{
ls_linkage_phases(input.seq, thres, twopt, mrk.to.add = NULL,
  prev.info = NULL)

\method{print}{two.pts.linkage.phases}(x, ...)

\method{plot}{two.pts.linkage.phases}(x, ...)
}
\arguments{
\item{input.seq}{an object of class \code{mappoly.sequence}.}

\item{thres}{the threshold used to determine if the linkage phases
compared via two-point analysis should be considered}

\item{twopt}{an object of class \code{poly.est.two.pts.pairwise}
containing the two-point information}

\item{mrk.to.add}{marker to be added at the end of the linkage
group. If \code{NULL} add all markers contained in
\code{mrk.seq}. Mostly for internal usage.}

\item{prev.info}{an object of class \code{two.pts.linkage.phases}
containing the previous info about linkage phase configuration.
Mostly for internal usage.}

\item{x}{an object of one of the classes \code{two.pts.linkage.phases}}

\item{...}{currently ignored}
}
\value{
An object of class \code{two.pts.linkage.phases} which
    contains a list of all possible linkage phase configuration in
    matrix form for both parents, P and Q.
}
\description{
Retunrs a list of possible linkage phase configurations using
mappoly.rf.matrix as a eliminations criteria
}
\examples{
  \dontrun{
    hexa_file<-system.file('extdata', 'hexa_fake', package = 'mappoly')
    hexa_dat<-read_geno(file_in = hexa_file)
    all_mrk<-make_seq_mappoly(hexa_dat, 'all')
    counts_all_mrk_from_web<-cache_counts_twopt(all_mrk,
                                                get.from.web=TRUE)
    all_pairs<-est_pairwise_rf(all_mrk, counts_all_mrk_from_web,
                               n.clusters=2)

    seq1<-make_seq_mappoly(hexa_dat, 'seq1')

    l_seq1_3.0 <- ls_linkage_phases(input.seq = seq1, thres = 3, twopt = all_pairs)
    l_seq1_3.0
    plot(l_seq1_3.0)
    l_seq1_0.5 <- ls_linkage_phases(input.seq = seq1, thres = 0.5, twopt = all_pairs)
    l_seq1_0.5
    plot(l_seq1_0.5)
    l_seq1_0.1 <- ls_linkage_phases(input.seq = seq1, thres = 0.1, twopt = all_pairs)
    l_seq1_0.1
    plot(l_seq1_0.1)
    }
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2018) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_. \url{https://doi.org/10.1101/415232}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
