% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_map_utils.R
\name{map_haplo}
\alias{map_haplo}
\title{States that should be visited given a molecular phenotype of the
haplotype}
\usage{
map_haplo(m, P, Q)
}
\arguments{
\item{m}{ploidy level}

\item{P}{a list of vectors indicating the haplotype configuration
for genitor P}

\item{Q}{a list of vectors indicating the haplotype configuration
for genitor Q}
}
\value{
a matrix whose the first two columns indicates all the
    possible states and the remaining columns indicates the
    observed molecular phenotype of the haplotype
}
\description{
States that should be visited given a molecular phenotype of the
haplotype
}
\keyword{internal}
