% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{mean_dist_from_truth}
\alias{mean_dist_from_truth}
\title{Calculates the mean of the square distances of points from their real position}
\usage{
mean_dist_from_truth(estmap, realmap)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Calculates the mean of the square distances of points from their real position
}
\keyword{internal}
