% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_datasets}
\alias{merge_datasets}
\title{Merge datasets}
\usage{
merge_datasets(dataset.1, dataset.2)
}
\arguments{
\item{dataset.1}{the first dataset of class \code{mappoly.data} to be merged}

\item{dataset.2}{the second dataset of class \code{mappoly.data} to be merged}
}
\value{
An object of class \code{mappoly.data} which contains all markers
from both datasets. It will be a list with the following components:
    \item{m}{ploidy level}
    \item{n.ind}{number individuals}
    \item{n.mrk}{total number of markers}
    \item{ind.names}{the names of the individuals}
    \item{mrk.names}{the names of the markers}
    \item{dosage.p}{a vector containing the dosage in
      parent P for all \code{n.mrk} markers}
    \item{dosage.q}{a vector containing the dosage in
      parent Q for all \code{n.mrk} markers}
    \item{sequence}{a vector indicating which sequence each marker
      belongs. Zero indicates that the marker was not assigned to any
      sequence}
    \item{sequence.pos}{Physical position of the markers into the
      sequence}
    \item{prob.thres}{(unused field)}
    \item{geno.dose}{a matrix containing the dosage for each markers (rows) 
      for each individual (columns). Missing data are represented by 
      \code{ploidy_level + 1}}
    \item{nphen}{(unused field)}
    \item{phen}{(unused field)}
    \item{chisq.pval}{a vector containing p-values related to the chi-squared 
    test of mendelian segregation performed for all markers}
}
\description{
This function merges two datasets of class \code{mappoly.data}. This can be useful
when individuals of a population were genotyped using two or more techniques
and have datasets in different files or formats. Please notice that the datasets
should contain the same number of individuals, which must be represented identically
in both datasets (e.g. \code{Ind_1} in both datasets, not \code{Ind_1}
in one dataset and \code{ind_1} or \code{Ind.1} in the other).
}
\details{
Please notice that if any of these datasets were generated using \code{read_vcf}
or \code{read_geno_dist} functions, the final dataset may not contain
some information, such as reference and alternative alleles,
marker depths and the probability distribution of genotypes.
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400378}
}
\author{
Gabriel Gesteira, \email{gabrielgesteira@usp.br}
}
