% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_map_by_steps.R
\name{merge_parental_maps}
\alias{merge_parental_maps}
\title{Build merged parental maps}
\usage{
merge_parental_maps(
  map.p1,
  map.p2,
  full.seq,
  full.mat,
  method = c("ols", "hmm"),
  verbose = TRUE
)
}
\arguments{
\item{map.p1}{object of class \code{mappoly.map} with parent 1 phased}

\item{map.p2}{object of class \code{mappoly.map} with parent 2 phased}

\item{full.seq}{object of class \code{mappoly.sequence} containing parent 1 and parent 2 markers}

\item{full.mat}{object of class \code{mappoly.rf.matrix} containing two-points recombination 
fraction estimations for parent 1 and parent 2 markers}

\item{method}{indicates whether to use 'hmm' (Hidden Markov Models), 'ols' (Ordinary Least Squares) 
to re-estimate the recombination fractions}
}
\value{
object of class \code{mappoly.map} with both parents information
}
\description{
Build merged parental maps
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu} with documentation and minor modifications by Cristiane Taniguti \email{chtaniguti@tamu.edu}
}
\keyword{internal}
