% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mrk_chisq_test}
\alias{mrk_chisq_test}
\title{Chi-square test}
\usage{
mrk_chisq_test(x, ploidy)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Chi-square test
}
\keyword{internal}
