% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_rf.R
\name{paralell_pairwise_discrete}
\alias{paralell_pairwise_discrete}
\title{Wrapper function to discrete-based pairwise two-point estimation in C++}
\usage{
paralell_pairwise_discrete(
  mrk.pairs,
  input.seq,
  geno,
  dP,
  dQ,
  count.cache,
  tol = .Machine$double.eps^0.25
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Wrapper function to discrete-based pairwise two-point estimation in C++
}
\keyword{internal}
