% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_rf.R
\name{paralell_pairwise_discrete_rcpp}
\alias{paralell_pairwise_discrete_rcpp}
\title{Wrapper function to discrete-based pairwise two-point estimation in C++}
\usage{
paralell_pairwise_discrete_rcpp(
  mrk.pairs,
  m,
  geno,
  dP,
  dQ,
  count.vector,
  count.phases,
  count.matrix.rownames,
  count.matrix.number,
  count.matrix.pos,
  count.matrix.length,
  tol = .Machine$double.eps^0.25,
  threads
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Wrapper function to discrete-based pairwise two-point estimation in C++
}
\keyword{internal}
