% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_rf.R
\name{paralell_pairwise_given_sparse_grid}
\alias{paralell_pairwise_given_sparse_grid}
\title{Wrapper function to discrete-based pairwise likelihood computation given rf = 1e-5}
\usage{
paralell_pairwise_given_sparse_grid(
  mrk.pairs,
  input.seq,
  geno,
  dP,
  dQ,
  count.cache
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Wrapper function to discrete-based pairwise likelihood computation given rf = 1e-5
}
\keyword{internal}
