% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_blocks.R
\name{parallel_block}
\alias{parallel_block}
\title{Auxiliary function to estimate a map in a block of markers using parallel 
processing}
\usage{
parallel_block(
  mrk.vec,
  dat.name,
  ph.thres = 3,
  tol = 0.01,
  phase.number.limit = 20,
  error = 0.05,
  tol.err = 0.001,
  verbose = FALSE
)
}
\description{
Auxiliary function to estimate a map in a block of markers using parallel 
processing
}
\keyword{internal}
