% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_marker_blocks.R
\name{parallel_hmm_bins}
\alias{parallel_hmm_bins}
\title{Wrap function to estimate a map using parallel processing (for find_marker_blocks)}
\usage{
parallel_hmm_bins(index.seq, thres.twopt = 10, thres.hmm = 3,
  extend.tail = 10, tol = 0.01, tol.final = 0.001,
  sub.map.size.limit = 7, phase.number.limit = 4, error = NULL,
  verbose = FALSE)
}
\arguments{
\item{index.seq}{sequence used as argument in function make_seq_temp}

\item{thres.twopt}{the threshold used to determine if the linakge
phases compared via two-point analysis should be considered}

\item{thres.hmm}{the threshold used to determine if the linakge
phases compared via hmm analysis should be considered}

\item{extend.tail}{trhe length of the tail of the chain that should
be used to calculate the likelihood of the linakge phases}

\item{tol}{the desired accuracy during the sequential phase.}

\item{tol.final}{the desired accuracy for the final map.}

\item{error}{global error rate}

\item{verbose}{print the progress}

\item{dat}{an object of class \code{mappoly.data}}

\item{dat.name}{name of the object of class \code{mappoly.data}}

\item{block.estimate}{logical. If TRUE returns a map forcing all
recombination fractions equal to 0 (1e-5)}
}
\value{
a list of maps
}
\description{
Wrap function to estimate a map using parallel processing (for find_marker_blocks)
}
\keyword{internal}
