% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phase_configurations.R
\name{ph_list_to_matrix}
\alias{ph_list_to_matrix}
\title{Linkage phase format conversion: list to matrix}
\usage{
ph_list_to_matrix(L, m)
}
\arguments{
\item{L}{a list of configuration phases}

\item{m}{ploidy level}
}
\value{
a matrix whose columns represent homologous chromosomes and
    the rows represent markers
}
\description{
This function converts linkage phase configurations from list
to matrix form
}
\keyword{internal}
