% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_phase_configurations.R
\name{ph_matrix_to_list}
\alias{ph_matrix_to_list}
\title{MAPpoly object conversion: phase configurations (matrix to list)}
\usage{
ph_matrix_to_list(M)
}
\arguments{
\item{M}{matrix whose columns represent homologous chromosomes and
the rows represent markers}
}
\value{
a list of configuration phases
}
\description{
This function converts linkage phase configurations from matrix
form to list
}
\keyword{internal}
