% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homolog_probs.R
\name{plot.mappoly.homoprob}
\alias{plot.mappoly.homoprob}
\title{Plots mappoly.homoprob}
\usage{
\method{plot}{mappoly.homoprob}(
  x,
  stack = FALSE,
  lg = NULL,
  ind = NULL,
  use.plotly = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{an object of class \code{mappoly.homoprob}}

\item{stack}{logical. If \code{TRUE}, probability profiles of all homologues
are stacked in the plot (default = FALSE)}

\item{lg}{indicates which linkage group should be plotted. If \code{NULL} 
(default), it plots the first linkage group. If 
\code{"all"}, it plots all linkage groups}

\item{ind}{indicates which individuals should be plotted. It can be the 
position of the individuals in the dataset or it's name. 
If \code{NULL} (default), the function plots the first 
individual}

\item{use.plotly}{if \code{TRUE} (default), it uses plotly interactive graphics}

\item{verbose}{if \code{TRUE} (default), the current progress is shown; if
\code{FALSE}, no output is produced}

\item{...}{unused arguments}
}
\description{
Plots mappoly.homoprob
}
