% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferential_pairing.R
\name{plot.mappoly.prefpair.profiles}
\alias{plot.mappoly.prefpair.profiles}
\title{Plots mappoly.prefpair.profiles}
\usage{
\method{plot}{mappoly.prefpair.profiles}(
  x,
  type = c("pair.configs", "hom.pairs"),
  min.y.prof = 0,
  max.y.prof = 1,
  thresh = 0.01,
  P1 = "P1",
  P2 = "P2",
  ...
)
}
\arguments{
\item{x}{an object of class \code{mappoly.prefpair.profiles}}

\item{type}{a character string indicating which type of graphic is plotted:
\code{"pair.configs"} (default) plots the preferential pairing 
profile for the pairing configurations or \code{"hom.pairs"} plots 
the preferential pairing profile for the homolog pairs}

\item{min.y.prof}{lower bound for y axis on the probability profile graphic (default = 0)}

\item{max.y.prof}{upper bound for y axis on the probability profile graphic (default = 1)}

\item{thresh}{threshold for chi-square test (default = 0.01)}

\item{P1}{a string containing the name of parent P1}

\item{P2}{a string containing the name of parent P2}

\item{...}{unused arguments}
}
\description{
Plots mappoly.prefpair.profiles
}
