% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{plot_diag_pc}
\alias{plot_diag_pc}
\title{Plot disgnostic graphics (2D)}
\usage{
plot_diag_pc(mappc, D1lim = NULL, D2lim = NULL, displaytext = TRUE)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Plot disgnostic graphics (2D)
}
\keyword{internal}
