% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{plot_diag_pc_3d}
\alias{plot_diag_pc_3d}
\title{Plot disgnostic graphics (3D)}
\usage{
plot_diag_pc_3d(mappc, D1lim = NULL, D2lim = NULL, D3lim = NULL,
  displaytext = TRUE)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Plot disgnostic graphics (3D)
}
\keyword{internal}
