% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genome_vs_map.R
\name{plot_genome_vs_map}
\alias{plot_genome_vs_map}
\title{Physical versus genetic distance}
\usage{
plot_genome_vs_map(map.list, config = "best")
}
\arguments{
\item{map.list}{A list or a single object of class \code{mappoly.map}}

\item{config}{A vector containing the position of the
configuration(s) to be plotted. If \code{'best'} (default), plots the configuration
with the highest likelihood for all elements in \code{'map.list'}}
}
\description{
This function plots scatterplot(s) of physical distance (in Mb) versus the genetic 
distance (in cM). Map(s) should be passed as a single object or a list of objects 
of class \code{mappoly.map}.
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400378}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
