% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_GIC}
\alias{plot_GIC}
\title{Genotypic information content}
\usage{
plot_GIC(hprobs, P = "P1", Q = "P2")
}
\arguments{
\item{hprobs}{an object of class \code{mappoly.homoprob}}

\item{P}{a string containing the name of parent P}

\item{Q}{a string containing the name of parent Q}
}
\description{
This function plots the genotypic information content given 
an object of class \code{mappoly.homoprob}.
}
\examples{
\donttest{
     w <- lapply(solcap.err.map[1:3], calc_genoprob)
     h.prob <- calc_homologprob(w)
     plot_GIC(h.prob)
}

}
