% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_map_by_steps.R
\name{plot_mappoly.map2}
\alias{plot_mappoly.map2}
\title{Plot object mappoly.map2}
\usage{
plot_mappoly.map2(x)
}
\arguments{
\item{x}{object of class \code{mappoly.map2}}
}
\description{
Plot object mappoly.map2
}
