% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_info.R
\name{plot_mrk_info}
\alias{plot_mrk_info}
\title{Plot marker information}
\usage{
plot_mrk_info(input.data, mrk.name)
}
\arguments{
\item{input.data}{an object of class \code{mappoly.data}}

\item{mrk.name}{marker name}
}
\description{
Plot summary statistics for a given marker
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2018) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_. \url{https://doi.org/10.1101/415232}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
