% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_list.R
\name{plot_one_map}
\alias{plot_one_map}
\title{plot a single linkage group with no phase}
\usage{
plot_one_map(x, i = 0, horiz = FALSE, col = "lightgray")
}
\description{
plot a single linkage group with no phase
}
\keyword{internal}
