% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_dist_hmm.R
\name{poly_hmm_est}
\alias{poly_hmm_est}
\title{Estimate genetic map using as input the probability distribution of
genotypes (wrapper function to C++)}
\usage{
poly_hmm_est(m, n.mrk, n.ind, p, dp, q, dq, g, rf, verbose = TRUE, tol = 0.001)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Estimate genetic map using as input the probability distribution of
genotypes (wrapper function to C++)
}
\keyword{internal}
