% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pos_twopt_est}
\alias{pos_twopt_est}
\title{Check if it is possible to estimate the recombination
fraction between neighbour markers using two-point
estimation}
\usage{
pos_twopt_est(input.seq)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Check if it is possible to estimate the recombination
fraction between neighbour markers using two-point
estimation
}
\keyword{internal}
