% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_mrk}
\alias{print_mrk}
\title{Summary of a set of markers}
\usage{
print_mrk(input.data, mrks)
}
\arguments{
\item{input.data}{an object \code{'mappoly.data'}}

\item{mrks}{marker sequence index (integer vector)}
}
\description{
Returns information related to a given set of markers
}
