% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polymap.R
\name{read_geno}
\alias{read_geno}
\alias{plot.mappoly.data}
\title{Data Input}
\usage{
read_geno(file.in)

\method{plot}{mappoly.data}(x, detailed = FALSE, ...)
}
\arguments{
\item{file.in}{the name of the input file which contains the data to
be read.}

\item{x}{an object of one of the classes \code{mappoly.data}}

\item{detailed}{if available, print the number of markers per sequence}

\item{...}{currently ignored}
}
\value{
An object of class \code{mappoly.data} which contains a
    list with the following components:
    \item{m}{ploidy level}
    \item{n.ind}{number individuals}
    \item{n.mrk}{total number of markers}
    \item{ind.names}{the names of the individuals}
    \item{mrk.names}{the names of the markers}
    \item{dosage.p}{a vector containing the dosage in
      parent P for all \code{n.mrk} markers}
    \item{dosage.q}{a vector containing the dosage in
      parent Q for all \code{n.mrk} markers}
    \item{sequence}{a vector indicating which sequence each marker
      belongs. Zero indicates that the marker was not assigned to any
      sequence.}
    \item{sequence.pos}{Physical position of the markers into the
      sequence}
    \item{geno.dose}{a matrix containing the dosage for each markers (rows) 
      for each individual (columns). Missing data are represented by 
      \code{ploidy_level + 1}.}
    \item{n.phen}{number of phenotypic traits}
    \item{phen}{a matrix containing the phenotypic data. The rows
                corespond to the trais and the columns correspond
                to the individuals}
}
\description{
Reads an external data file. The format of the file is described in \code{details}
section. This fucntion creates an object of class \code{mappoly.data}
}
\details{
The first line of the input file contains the word \code{ploidy}
followed by the ploidy level of the parents.  The second and third
lines contains the words \code{nind} and \code{nmrk} followed by
the number of individuals in the dataset and the total number of
markers, respectively. The lines number 4 and 5 should contain the
word \code{mrknames} and \code{indnames} followed by a sequence of
the names of the markers and the name of the individuals,
respectively. Lines 6 and 7 contain the words \code{dosageP} and
\code{dosageQ} followed by a sequence of numbers containing the
dosage of all markers in parent 1 and 2. Line 8, contain the string
\code{seq} fallowed by a sequence of integer numbers indicating the
sequence each marker belongs. It can be any 'a priori' information
regarding the physical distance between markers. For example, these
numbers could refer to chromosomes, scaffolds or even contigs, in
which the markers are positioned. If this information is not
available for a particular marker, NA should be used. If this
information is not available for any of the markers, the string
\code{seq} should be followed by a single NA. Line number 9 contain
the word \code{seqpos} fallowed by the physical position of the
markers into the sequence. The physical position can be given in
any unity of physical genomic distance (base pairs, for
instance). However, the user should be able to take decisions based
on these values, such as occurrence of crossing overs. Line number
10 should contains the word \code{nphen} fallowed by the number of
phenotypic traits. Line number 11 is skipped (Usually used as
spacer).  The next elements are strings containing the name of the
phenotypic trait followed by the phenotypic values. The number of
lines should be the same number of phenotypic traits. Missing
values are represented by \code{NA}. The line number 12 +
\code{nphen} is skipped.  Finally, the last element is a table
containing the dosage for each markers (rows) for each individual
(columns). Missing data are represented by \code{NA}.'
}
\examples{
\dontrun{
    hexa.file<-system.file('extdata', 'hexafake', package = 'mappoly')
    hexafake<-read_geno(file.in  = hexa.file)
    print(hexafake, detailed = TRUE)

    ## Same thing
    data(hexafake)
}
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2018) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_. \url{https://doi.org/10.1101/415232}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
