% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mappoly_csv.R
\name{read_geno_csv}
\alias{read_geno_csv}
\title{Data Input in CSV format}
\usage{
read_geno_csv(file.in, ploidy, filter.non.conforming = TRUE)
}
\arguments{
\item{file.in}{a character string with the name of (or full path to) the input file which contains the data to
be read}

\item{ploidy}{the ploidy level}

\item{filter.non.conforming}{if \code{TRUE} (default) exclude samples with non 
expected genotypes under random chromosome pairing and no double reduction}
}
\value{
An object of class \code{mappoly.data} which contains a
    list with the following components:
    \item{m}{ploidy level}
    \item{n.ind}{number individuals}
    \item{n.mrk}{total number of markers}
    \item{ind.names}{the names of the individuals}
    \item{mrk.names}{the names of the markers}
    \item{dosage.p}{a vector containing the dosage in
      parent P for all \code{n.mrk} markers}
    \item{dosage.q}{a vector containing the dosage in
      parent Q for all \code{n.mrk} markers}
    \item{sequence}{a vector indicating which sequence each marker
      belongs. Zero indicates that the marker was not assigned to any
      sequence}
    \item{sequence.pos}{Physical position of the markers into the
      sequence}
    \item{geno.dose}{a matrix containing the dosage for each markers (rows) 
      for each individual (columns). Missing data are represented by 
      \code{ploidy_level + 1}}
    \item{n.phen}{number of phenotypic traits}
    \item{phen}{a matrix containing the phenotypic data. The rows
                corespond to the trais and the columns correspond
                to the individuals}
}
\description{
Reads an external comma-separated values (CSV) data file. The format of the file is described in the \code{Details}
section. This function creates an object of class \code{mappoly.data}.
}
\details{
This is an alternative and a somewhat more straightforward version of the function 
 \code{\link[mappoly]{read_geno}}. The input is a standard CSV file where the rows 
 represent the markers, except for the first row which is used as a header. 
 The first five columns contain the marker names, the dosage in parents 1 and 2, 
 the sequence information (i.e. chromosome,  scaffold, contig, etc) and the 
 position of the marker within the sequence. The remaining columns contain 
 the dosage of the full-sib population. A tetraploid example of such file 
 can be found in \code{ins/extdata/tetra_solcap.csv}
}
\examples{
\dontrun{
    solcap.file <- system.file('extdata', 'tetra_solcap.csv', package = 'mappoly')
    tetra.solcap <- read_geno_csv(file.in  = solcap.file, ploidy = 4)
    print(tetra.solcap, detailed = TRUE)

    ## Same thing
    data("tetra.solcap")
}
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400378}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
