% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{recalc_nnfit_from_map}
\alias{recalc_nnfit_from_map}
\title{Calculates a new nnfit based on a new order after estimating a map.}
\usage{
recalc_nnfit_from_map(estmap, mapobject, fname = NULL)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Calculates a new nnfit based on a new order after estimating a map.
}
\keyword{internal}
