% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reest_map_geno_dist.R
\name{reest_map_geno_dist}
\alias{reest_map_geno_dist}
\title{Reestimate map using genotype distribution}
\usage{
reest_map_geno_dist(input.map, dat.dist, phase.config = "best",
  verbose = TRUE, tol = 0.01)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Reestimate map using genotype distribution
}
\keyword{internal}
