% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reestimate_map.R
\name{reest_rf}
\alias{reest_rf}
\title{Reestimate the recombination fractions in a genetic map}
\usage{
reest_rf(input.map, input.mat = NULL, tol = 0.01,
  phase.config = "all", method = c("hmm", "ols"), weight = TRUE,
  verbose = TRUE)
}
\arguments{
\item{input.map}{An object of class \code{mappoly.map}}

\item{input.mat}{An object of class \code{mappoly.rf.matrix}}

\item{tol}{tolerance for determining convergence.}

\item{phase.config}{should be a string \code{'best'} or the position of the
configuration to be plotted. If \code{'best'}, plot the configuration
with the highest likelihood.}

\item{method}{indicates whether to use Hidden Markov Models or Ordinary
Least Squares to reestimate the recombination fraction}

\item{weight}{if \code{TRUE}, it uses the LOD scores to perform a weighted
regression when the Ordinary Least Saquares is chosen}

\item{verbose}{if \code{TRUE}, current progress is shown; if
\code{FALSE}, no output is produced.}
}
\value{
a numeric vector of size \code{m} indicating which
    homologous in h2 represents the homologous in h1. If there is
    no correspondence, i.e. different homologous, it returns NA for
    that homologous.
}
\description{
Reestimate the recombination fractions in a genetic map
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2018) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_. \url{https://doi.org/10.1101/415232}
}
