% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_marker}
\alias{remove_marker}
\title{Remove markers from a sequence}
\usage{
remove_marker(input.seq, mrk)
}
\arguments{
\item{input.seq}{an object of class \code{"mappoly.sequence"}}

\item{mrk}{markers to be removed from the input sequence, identified by its positions on dataset. For multiple markers, please use 'c()'.}
}
\value{
An object of class \code{mappoly.sequence}, which is a
    list containing the following components:
    \item{seq.num}{a \code{vector} containing the (ordered) indices
        of markers in the sequence, according to the input file.}
    \item{seq.phases}{a \code{list} with the linkage phases between
        markers in the sequence, in corresponding positions. \code{-1}
        means that there are no defined linkage phases.}
    \item{seq.rf}{a \code{vector} with the recombination
        frequencies between markers in the sequence. \code{-1} means
        that there are no estimated recombination frequencies.}
    \item{loglike}{log-likelihood of the corresponding linkage
        map.}
    \item{data.name}{name of the object of class
        \code{mappoly.data} with the raw data.}
    \item{twopt}{name of the object of class \code{mappoly.twopt}
        with the 2-point analyses. \code{-1} means that the twopt
        estimates were not computed.}
}
\description{
This function creates a new sequence by removing markers from an existing sequence.
}
\examples{
\dontrun{
data(hexafake)
seq1.mrk<-make_seq_mappoly(hexafake, 'seq1')

# Removing marker 1
seq1.m1.mrk<-remove_marker(seq1.mrk, mrk = 1)

# Removing markers 1, 2 and 3
seq1.mm.mrk<-remove_marker(seq1.mrk, mrk = c(1,2,3))
}

}
\author{
Gabriel Gesteira and Marcelo Mollinari, \email{gabrielgesteira@usp.br}
}
