% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_list_to_matrix.R
\name{rf_list_to_matrix}
\alias{rf_list_to_matrix}
\title{Recombination fraction list to matrix}
\usage{
rf_list_to_matrix(input.twopt, thresh.LOD.ph = 0, thresh.LOD.rf = 0,
  thresh.rf = 0.5, n.clusters = 1, verbose = TRUE)
}
\arguments{
\item{input.twopt}{an object of class \code{poly.est.two.pts.pairwise}}

\item{thresh.LOD.ph}{LOD score threshold for linkage phase
configuration.}

\item{thresh.LOD.rf}{LOD score threshold for recombination phase}

\item{thresh.rf}{recombination fraction threshold}

\item{n.clusters}{Number of parallel processes to spawn}

\item{verbose}{if \code{TRUE}, current progress is shown; if
\code{FALSE}, no output is produced.}

\item{x}{an object of class mappoly.rf.matrix}

\item{type}{type of matrix that shuold be printed. Can be one of the
following: \code{"rf"}, for recombination fraction or \code{"lod"}
for LOD Score}

\item{ord}{the order in which the markes should be ploted}

\item{rem}{which markers should be removed from the heatmap}

\item{main.text}{title of the heatmap}

\item{index}{should the numbers corresponding to the markers be printed in the diagonal of the heatmap?}

\item{...}{currently ignored}
}
\value{
A list containing two matrices. The first one contains the
    filtered recombination fraction and the second one contains the
    information matrix
}
\description{
Transforms the recombination fraction list contained in an object
of class \code{poly.est.two.pts.pairwise} into a recombination
fraction matrix
}
\details{
\code{thresh_LOD_ph} should be set in order to only selects
    recombination fractions which have LOD scores associated to the
    linkage phase configuration bigger than \code{thresh_LOD_ph}
    for the second most likely linkage phase configuration.
}
\examples{
  \dontrun{
    data(hexafake)
    all.mrk<-make_seq_mappoly(hexafake, 'all')
    red.mrk<-elim_redundant(all.mrk)
    unique.mrks<-make_seq_mappoly(red.mrk)
    counts.web<-cache_counts_twopt(unique.mrks, get.from.web = TRUE)
    all.pairs<-est_pairwise_rf(input.seq = unique.mrks,
                               count.cache = counts.web,
                               n.clusters = 16,
                               verbose=TRUE)

    ## Full recombination fraction matrix
    mat.full<-rf_list_to_matrix(input.twopt=all.pairs)
    plot(mat.full)
    plot(mat.full, type = "lod")

    ## Fintered matrix
    mat.filt<-rf_list_to_matrix(input.twopt=all.pairs,
                                thresh.LOD.ph = 5,
                                thresh.LOD.rf = 5,
                                thresh.rf = 0.5,
                                n.clusters = 1,
                                verbose = TRUE)
    plot(mat.filt)
    plot(mat.filt, type = "lod")
 }
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2017) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
