% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg_poly.R
\name{segreg_poly}
\alias{segreg_poly}
\title{Polysomic segregation frequency}
\usage{
segreg_poly(m, dP, dQ)
}
\arguments{
\item{m}{ploidy level}

\item{dP}{dosage in parent P}

\item{dQ}{dosage in parent Q}
}
\value{
a vector containing the expected segregation frequency for
    all possible genotypic classes.
}
\description{
Computes the polysomic segregation frequency given a ploidy level
and the dosage of the locus in both parents.
}
\examples{

#auto-octaploid with two and three doses in parents P and Q,
#respectively
(seg<-segreg_poly(m=8, dP=2, dQ=3))
barplot(seg, las=2)

}
\references{
Mollinari, M., and Garcia, A.  A. F. (2017) Theoretical
    Basis of Genetic Mapping in Complex Polyploids Using
    Hidden Markov Models, _submited_
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
