% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_list_to_matrix.R
\name{select_rf}
\alias{select_rf}
\title{Select rf adn lod based on thresholds}
\usage{
select_rf(x, thresh.LOD.ph, thresh.LOD.rf, thresh.rf, shared.alleles = FALSE)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Select rf adn lod based on thresholds
}
\keyword{internal}
