% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utils.R
\name{sim_cross_one_informative_parent}
\alias{sim_cross_one_informative_parent}
\title{Simulate mapping population (one parent)}
\usage{
sim_cross_one_informative_parent(m, n.mrk, rf.vec, hom.allele, n.ind,
  seed = NULL, prob = NULL)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
This function simulates a polyploid mapping population
under random chromosome segregation
with one informative parent. This function is not to be
directly called by the user
}
\examples{
 \dontrun{
 sim_cross_one_informative_parent(m=6,
                                  n.mrk=2,
                                  rf.vec=.1,
                                  hom.allele=list(c(2,4), c(2,3,6)),
                                  n.ind=10)

}
}
\keyword{internal}
