% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utils.R
\name{sim_cross_two_informative_parents}
\alias{sim_cross_two_informative_parents}
\title{Simulate mapping population (tow parents)}
\usage{
sim_cross_two_informative_parents(
  m,
  n.mrk,
  rf.vec,
  n.ind,
  hom.allele.p,
  hom.allele.q,
  prob.P = NULL,
  prob.Q = NULL,
  seed = NULL
)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Simulate mapping population (tow parents)
}
\examples{
 \dontrun{
hom.allele.p<-list(c(1,3,4), c(2,3,4), c(1,4,5,6), c(1,3,4,7), 1)
hom.allele.q<-list(c(2,4), c(2,3,6), c(1,4,5,8), 0, c(1,2))
x<-sim_cross_two_informative_parents(m=8,
                                     n.mrk=5,
                                     rf.vec=.1,
                                     n.ind=100,
                                     hom.allele.p,
                                     hom.allele.q)
                                     }
}
\keyword{internal}
