% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_homologous.R
\name{sim_homologous}
\alias{sim_homologous}
\title{Simulate homology groups}
\usage{
sim_homologous(m, n.mrk, min.d = 0, max.d, prob.dose = NULL, max.ph,
  restriction = TRUE, seed = NULL)
}
\arguments{
\item{m}{ploidy level. Must be an even number.}

\item{n.mrk}{number of markers.}

\item{min.d}{minimum dosage to be simulated}

\item{max.d}{maximum dosage to be simulated}

\item{prob.dose}{a vector indicating the proportion of markers for
different dosage to be simulated. (currently ignored)}

\item{max.ph}{maximum phase difference. It is ignored if
\code{choose.hom.at.random=TRUE}}

\item{restriction}{if TRUE, avoid cases where it is impossible to
estmate recombination fraction and/or linkage phases via
two-point analysis.}

\item{seed}{random number generator seed}
}
\value{
S3
A list which is a list containing the following components:
  \item{hom.allele.p}{ a \code{list} of vectors
   containing linkage phase configuration. Each vector contains the
   numbers of the homologous chromosomes in which the alleles are
   located. For instance, a vector containing \eqn{(1,3,4)} means that
   the marker has three doses located in the chromosomes 1, 3 and 4. For
   zero doses, use 0.}
  \item{p}{contains the indices of the starting positions of the
    dosages, considering that the vectors contained in \code{p} are
    concatenated. Markers with no doses (zero doses are also
    considered).}
  \item{hom.allele.q}{Analogously to hom.allele.p}
  \item{q}{Analogously to p}
}
\description{
Simulate two homology groups (one for each parent) and their
linkage phase configuration.
}
\details{
This function prevents the simulation of linkage phase
configurations which are impossible to estimate via two point
methods
}
\examples{
    h.temp<-sim_homologous(m=6, n.mrk=20, max.d=3, max.ph=3,
                           seed=123)

}
\references{
Mollinari, M., and Garcia, A.  A. F. (2018) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_. \url{https://doi.org/10.1101/415232}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
