% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solcap.maps}
\alias{solcap.maps}
\alias{solcap.dose.map}
\alias{solcap.map}
\alias{solcap.mds.map}
\alias{solcap.prior.map}
\alias{solcap.err.map}
\title{Solcap Maps

Resulting maps from \code{\link[mappoly]{tetra.solcap}} and \code{\link[mappoly]{tetra.solcap.geno.dist}} datasets.}
\format{All maps have the same format: a list containing 12 objects of class \code{mappoly.map}, each one 
representing one linkage group. 
\describe{
    \item{solcap.map}{genetic map using genomic order and dosage call}
    \item{solcap.mds.map}{genetic map using \code{\link[mappoly]{mds_mappoly} order and dosage call}}
    \item{solcap.prior.map}{genetic map using genomic order and prior probability distribution}
    \item{solcap.err.map}{genetic map using genomic order, dosage call and global calling error}     
}}
\usage{
solcap.dose.map

solcap.mds.map

solcap.prior.map

solcap.err.map
}
\description{
Solcap Maps

Resulting maps from \code{\link[mappoly]{tetra.solcap}} and \code{\link[mappoly]{tetra.solcap.geno.dist}} datasets.
}
\keyword{datasets}
