% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_parallel.R
\name{submap_merge}
\alias{submap_merge}
\title{Function to merge submaps two-by-two}
\usage{
submap_merge(submaps, tpt, n.cores, platform, thres.twopt2, thres.hmm2, tol)
}
\arguments{
\item{submaps}{a list of \code{mappoly.map} objects}

\item{tpt}{a two points object}

\item{n.cores}{number of cores}

\item{platform}{automatically detected}

\item{thres.twopt2}{two-point LOD threshold to eliminate phases}

\item{thres.hmm2}{LOD threshold to eliminate map phases}

\item{tol}{precision for hmm convergence}
}
\description{
Function to merge submaps two-by-two
}
\author{
Gabriel Gesteira, \email{gabrielgesteira@usp.br}
}
\keyword{internal}
