% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mappoly_csv.R
\name{table_to_mappoly}
\alias{table_to_mappoly}
\title{Conversion of data.frame to mappoly.data}
\usage{
table_to_mappoly(
  dat,
  ploidy,
  filter.non.conforming = TRUE,
  elim.redundant = TRUE,
  verbose = TRUE
)
}
\description{
Conversion of data.frame to mappoly.data
}
\keyword{internal}
