% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_map}
\alias{update_map}
\title{Update map}
\usage{
update_map(map)
}
\arguments{
\item{map}{an map object of class \code{mappoly.map}}
}
\value{
an updated object of class \code{mappoly.map}, containing the original map plus redundant markers
}
\description{
This function takes an object of class \code{mappoly.map} and checks for
removed redundant markers in the original dataset. Once redundant markers
are found, they are re-added to the map in their respective equivalent positions
and another HMM round is performed.
}
\author{
Gabriel Gesteira, \email{gabrielgesteira@usp.br}
}
