% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_missing}
\alias{update_missing}
\title{Update missing information}
\usage{
update_missing(input.data, prob.thres = 0.95)
}
\arguments{
\item{input.data}{an object of class \code{mappoly.data}}

\item{prob.thres}{probability threshold to associate a marker call to a 
dosage. Markers with maximum genotype probability smaller than 'prob.thres' 
are considered as missing data for the dosage calling purposes}
}
\description{
Updates the missing data in the dosage matrix of an object of class 
\code{mappoly.data} given a new probability threshold
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2018) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_. \url{https://doi.org/10.1101/415232}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
