% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_hmm.R
\name{update_ph_list_at_hmm_thres}
\alias{update_ph_list_at_hmm_thres}
\title{makes a phase list from map, selecting only 
configurations under a certain threshold}
\usage{
update_ph_list_at_hmm_thres(map, thres.hmm)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
makes a phase list from map, selecting only 
configurations under a certain threshold
}
\keyword{internal}
