% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_marker}
\alias{add_marker}
\title{Add a single marker to a map}
\usage{
add_marker(input.map, mrk, pos, rf.matrix, genoprob = NULL,
  phase.config = "best", tol = 0.001)
}
\arguments{
\item{input.map}{an object of class \code{"mappoly.map"}}

\item{mrk}{the name of the marker to be inserted}

\item{pos}{the name of the marker after which the new marker should be added.
One also can inform the numeric position (between markers) were the 
new marker should be added. To insert a marker at the beginning of a 
map, one should use \code{pos = 0}.}

\item{rf.matrix}{an object of class \code{"mappoly.rf.matrix"}}

\item{genoprob}{an object of class \code{"mappoly.genoprob"}}

\item{phase.config}{which phase configuration should be used. "best" (default) 
will choose the maximum likelihood configuration.}

\item{tol}{the desired accuracy.}
}
\value{
An object of class 'mappoly.map'
}
\description{
Creates a new map by adding a marker in a given position in a pre-built map.
}
\details{
\code{add_marker} splits the input map in two submaps to the left and to the right 
of the given position, and using the genotype probabilities, computes the log-likelihhod
of all possible linkage phases under a two-point threshold.
}
\examples{
\dontrun{
input.map <- maps.hexafake[[1]]
mrk <- c(4, 9, 23)
mrk <- c("M_85", "M_152")
output.map <- drop_marker(input.map, mrk)
plot(output.map, mrk.names = TRUE, cex = .5)
}

}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
