% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_genoprob.R
\name{calc_genoprob}
\alias{calc_genoprob}
\title{Compute genotype conditional probabilities}
\usage{
calc_genoprob(input.map, phase.config = "best", verbose = TRUE)
}
\arguments{
\item{input.map}{An object of class \code{mappoly.map}}

\item{phase.config}{which phase configuration should be used. "best" (default) 
will choose the phase configuration associated with the
maximum likelihood}

\item{verbose}{if \code{TRUE} (default), current progress is shown; if
\code{FALSE}, no output is produced}

\item{...}{currently ignored}
}
\value{
An object of class 'mappoly.genoprob' which has two elements: a tridimensional
array containing the probabilities of all possible genotypes for each individual
in each marker position; and the marker sequence with it's recombination frequencies
}
\description{
Conditional genotype probabilities are calculated for each marker
position and each individual given a map. In this version,
the probabilities are not calculated between markers.
}
\examples{
 \dontrun{
    data(hexafake)
    mrk.subset<-make_seq_mappoly(hexafake, 1:100)
    red.mrk<-elim_redundant(mrk.subset)
    unique.mrks<-make_seq_mappoly(red.mrk)
    counts.web<-cache_counts_twopt(unique.mrks, get.from.web = TRUE)
    subset.pairs<-est_pairwise_rf(input.seq = unique.mrks,
                                  count.cache = counts.web,
                                  n.clusters = 16,
                                  verbose=TRUE)
    system.time(
    subset.map <- est_rf_hmm_sequential(input.seq = unique.mrks,
                                        thres.twopt = 5,
                                        thres.hmm = 3,
                                        extend.tail = 50,
                                        tol = 0.1,
                                        tol.final = 10e-3,
                                        twopt = subset.pairs,
                                        verbose = TRUE))
   probs<-calc_genoprob(input.map = subset.map,
                        verbose = TRUE)
   image(t(probs$probs[,,1]), col = rev(heat.colors(100)))
 }

}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400378}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
