% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_map_utils.R
\name{calc_genoprob_haplo}
\alias{calc_genoprob_haplo}
\title{Compute conditional probabilities of the genotypes given a sequence 
of block markers}
\usage{
calc_genoprob_haplo(
  ploidy,
  n.mrk,
  n.ind,
  haplo,
  emit = NULL,
  rf_vec,
  ind.names,
  verbose = TRUE,
  highprec = FALSE
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Compute conditional probabilities of the genotypes given a sequence 
of block markers
}
\keyword{internal}
