% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homolog_probs.R
\name{calc_homoprob}
\alias{calc_homoprob}
\alias{print.mappoly.homoprob}
\alias{plot.mappoly.homoprob}
\title{Compute homolog probabilities}
\usage{
calc_homoprob(input.probs)

\method{print}{mappoly.homoprob}(x, ...)

\method{plot}{mappoly.homoprob}(x, ind = NULL, use.plotly = TRUE, ...)
}
\arguments{
\item{input.probs}{an object of class \code{"mappoly.genoprob"}}

\item{x}{an object of class \code{"mappoly.homoprob"}}

\item{...}{unused argument}

\item{ind}{indicates which individual should be plotted. It can be the 
position of the individuals in the data set or its name. 
If \code{NULL} (default), the function plots the first 
individual.}

\item{use.plotly}{if \code{TRUE}, it uses plotly interactive graphics}
}
\description{
Compute homolog probabilities for all individuals in the full-sib
offspring
}
\examples{
\dontrun{
  ## hexaploid example
  w<-calc_genoprob(maps.hexafake[[1]])
  h.prob.ch1<-calc_homoprob(w)
  print(h.prob.ch1)
  plot(h.prob.ch1)
  ## tetraploid solcap example
  #load("~/repos/MAPpoly_vignettes/vignette_1/maps.rda")
  h.prob.solcap.ch1<-calc_homoprob(genoprob.err[[1]])
  print(h.prob.solcap.ch1)
  plot(h.prob.solcap.ch1)
}
}
\references{
Mollinari, M. et al. (2019) Unraveling the hexaploid 
    sweetpotato inheritance using ultra-dense multilocus mapping, 
    _submited_. \url{https://doi.org/10.1101/689638}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
