% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homolog_probs.R
\name{calc_homoprob}
\alias{calc_homoprob}
\alias{print.mappoly.homoprob}
\alias{plot.mappoly.homoprob}
\title{Homolog probabilities}
\usage{
calc_homoprob(input.genoprobs)

\method{print}{mappoly.homoprob}(x, ...)

\method{plot}{mappoly.homoprob}(x, stack = FALSE, lg = NULL,
  ind = NULL, use.plotly = TRUE, ...)
}
\arguments{
\item{input.genoprobs}{an object of class \code{"mappoly.genoprob"}}

\item{x}{an object of class \code{"mappoly.homoprob"}}

\item{...}{unused arguments}

\item{stack}{loagical. If \code{TRUE}, probability profiles of all homologs
are stacked in the plot.}

\item{lg}{indicates which linkage group should be plotted. If \code{NULL} 
(default), the function plots the first linkage group. If 
\code{"all"}, the function plots all linkage groups.}

\item{ind}{indicates which individual should be plotted. It can be the 
position of the individuals in the data set or its name. 
If \code{NULL} (default), the function plots the first 
individual.}

\item{use.plotly}{if \code{TRUE}, it uses plotly interactive graphics}
}
\description{
Compute homolog probabilities for all individuals in the full-sib
population
}
\examples{
\dontrun{
  ## hexaploid example
  w1 <- lapply(maps.hexafake, calc_genoprob)
  h.prob <- calc_homoprob(w1)
  print(h.prob)
  plot(h.prob)
  plot(h.prob, lg = 1, ind = 5, use.plotly = FALSE)
  plot(h.prob, lg = c(1,3), ind = 15, use.plotly = FALSE)
  plot(h.prob, lg = "all")
  
  ## tetraploid solcap example
  w2<-lapply(solcap.dose.map, calc_genoprob)
  h.prob.solcap<-calc_homoprob(w2)
  print(h.prob.solcap)
  plot(h.prob.solcap, ind = "ind_10")
  plot(h.prob.solcap, stack = TRUE, ind = 5)
  
  w3<-lapply(solcap.err.map, calc_genoprob_error, error = 0.05)
  h.prob.solcap.err<-calc_homoprob(w3)
  plot(h.prob.solcap, lg = 1, ind = 100, use.plotly = FALSE)
  plot(h.prob.solcap.err, lg = 1, ind = 100, use.plotly = FALSE)
}

}
\references{
Mollinari, M. et al. (2019) Unraveling the hexaploid 
    sweetpotato inheritance using ultra-dense multilocus mapping, 
    _submited_. \url{https://doi.org/10.1101/689638}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
