% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homolog_probs.R
\name{calc_homoprob}
\alias{calc_homoprob}
\title{Homolog probabilities}
\usage{
calc_homoprob(input.genoprobs)
}
\arguments{
\item{input.genoprobs}{an object of class \code{mappoly.genoprob}}
}
\description{
Compute homolog probabilities for all individuals in the full-sib
population given a map and conditional genotype probabilities.
}
\examples{
\dontrun{
  ## tetraploid solcap example
  w2<-lapply(solcap.dose.map, calc_genoprob)
  h.prob.solcap<-calc_homoprob(w2)
  print(h.prob.solcap)
  plot(h.prob.solcap, ind = "ind_10")
  plot(h.prob.solcap, stack = TRUE, ind = 5)
  plot(h.prob.solcap, stack = TRUE, ind = 5, lg = "all")
  
  w3<-lapply(solcap.err.map, calc_genoprob_error, error = 0.05)
  h.prob.solcap.err<-calc_homoprob(w3)
  plot(h.prob.solcap, lg = 1, ind = 100, use.plotly = FALSE)
  plot(h.prob.solcap.err, lg = 1, ind = 100, use.plotly = FALSE)
  
  ## hexaploid example
  w1 <- lapply(maps.hexafake, calc_genoprob)
  h.prob <- calc_homoprob(w1)
  print(h.prob)
  plot(h.prob)
  plot(h.prob, lg = 1, ind = 5, use.plotly = FALSE)
  plot(h.prob, lg = c(1,3), ind = 15, use.plotly = FALSE)
  plot(h.prob, lg = "all")
  
}

}
\references{
Mollinari M., Olukolu B. A.,  Pereira G. da S., 
    Khan A., Gemenet D., Yench G. C., Zeng Z-B. (2020), 
    Unraveling the Hexaploid Sweetpotato Inheritance 
    Using Ultra-Dense Multilocus Mapping, 
    _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400620}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
