% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferential_pairing.R
\name{calc_prefpair_profiles}
\alias{calc_prefpair_profiles}
\alias{print.mappoly.prefpair.profiles}
\alias{plot.mappoly.prefpair.profiles}
\title{Preferential pairing profiles}
\usage{
calc_prefpair_profiles(input.genoprobs)

\method{print}{mappoly.prefpair.profiles}(x, ...)

\method{plot}{mappoly.prefpair.profiles}(x, min.y.prof = 0,
  max.y.prof = 1, thresh = 0.01, ...)
}
\arguments{
\item{input.genoprobs}{an object of class \code{"mappoly.genoprob"}}

\item{x}{an object of class \code{"mappoly.prefpair.profiles"}}

\item{...}{unused arguments}

\item{min.y.prof}{lower bound for y axis on the probability profile graphic.}

\item{max.y.prof}{upper bound for y axis on the probability profile graphic.}

\item{thresh}{line for chi-square test}
}
\description{
Compute the probability profiles for all pairing configurations in both parents.
}
\examples{
\dontrun{
  ## hexaploid example
  w1 <- lapply(maps.hexafake, calc_genoprob)
  x1 <- calc_prefpair_profiles(w1)
  print(x1)
  plot(x1, min.y.prof = 0.05, max.y.prof = .15, thresh = 0.01)
  
  ## tetraploid example
  w2 <- lapply(solcap.err.map, calc_genoprob_error, error = 0.05)
  x2 <- calc_prefpair_profiles(w2)
  print(x2)
  plot(x2)
}

}
\references{
Mollinari, M. et al. (2019) Unraveling the hexaploid 
    sweetpotato inheritance using ultra-dense multilocus mapping, 
    _submited_. \url{https://doi.org/10.1101/689638}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
