% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_data_sanity}
\alias{check_data_sanity}
\title{Data sanity check}
\usage{
check_data_sanity(x)
}
\arguments{
\item{x}{an object of class \code{mappoly.data}}
}
\description{
Checks the consistency of a dataset
}
\examples{
\dontrun{

    solcap.dose.file <- system.file('extdata', 'tetra_solcap_geno', package = 'mappoly')
    dat.dose <- read_geno(file.in  = solcap.dose.file)
    check_data_sanity(dat.dose)

    solcap.file <- system.file('extdata', 'tetra_solcap_geno_dist.bz2', package = 'mappoly')
    dat.dist <- read_geno_dist(file.in  = solcap.file)
    check_data_sanity(dat.dist)
}
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400378}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
