% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compare_haplotypes}
\alias{compare_haplotypes}
\title{Compare two polyploid haplotypes stored in list format}
\usage{
compare_haplotypes(m, h1, h2)
}
\arguments{
\item{m}{ploidy level}

\item{h1}{homology group 1}

\item{h2}{homology group 2}
}
\value{
a numeric vector of size \code{m} indicating which
    homolog in h2 represents the homolog in h1. If there is
    no correspondence, i.e. different homolog, it returns NA for
    that homolog.
}
\description{
Compare two polyploid haplotypes stored in list format
}
\keyword{internal}
