% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_genoprob.R
\name{create_map}
\alias{create_map}
\title{Create a map with pseudomarkers at a given step}
\usage{
create_map(input.map, step = 0, phase.config = "best")
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Create a map with pseudomarkers at a given step
}
\keyword{internal}
