% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct_map_with_genotyping_error.R
\name{est_full_hmm_with_global_error}
\alias{est_full_hmm_with_global_error}
\title{Reestimate gentic map given a global genotyping error}
\usage{
est_full_hmm_with_global_error(input.map, error = NULL, tol = 0.001,
  th.prob = 0.95, verbose = FALSE)
}
\arguments{
\item{input.map}{an object of class \code{mappoly.map}.}

\item{error}{global error rate}

\item{tol}{the desired accuracy.}

\item{th.prob}{the threshold for using global error or genotype 
probability distribution contained in the data set}

\item{verbose}{if \code{TRUE}, current progress is shown; if
\code{FALSE}, no output is produced.}
}
\value{
An object of class 'mappoly.map'
}
\description{
This function considers a global error when reestimating
a genetic map using hdden Markov models
}
\examples{
  \dontrun{
    hexa.file<-system.file('extdata', 'hexafake', package = 'mappoly')
    hexa.dat<-read_geno(file.in = hexa.file)
    seq1.20<-make_seq_mappoly(hexa.dat, 1:20)
    counts<-cache_counts_twopt(seq1.20,
                               get.from.web=TRUE)
    subset.pairs<-est_pairwise_rf(seq1.20, counts,
                               n.clusters=1)
    subset.map <- est_rf_hmm_sequential(input.seq  = seq1.20,
                                        thres.twopt = 5,
                                        thres.hmm = 10,
                                        extend.tail = 10,
                                        tol = 0.1,
                                        tol.final = 10e-4,
                                        twopt = subset.pairs,
                                        verbose = TRUE)
    subset.map
    plot(subset.map)
     
    subset.map.reest<-est_full_hmm_with_global_error(subset.map, 
                                                     error=0.01, 
                                                     tol=10e-4, 
                                                     verbose = TRUE)
    subset.map.reest
    plot(subset.map.reest)
    }

}
\references{
Mollinari, M., and Garcia, A.  A. F. (2018) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_. \url{https://doi.org/10.1101/415232}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
