% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reestimate_map_with_global_genotypic_error.R
\name{est_full_hmm_with_global_error}
\alias{est_full_hmm_with_global_error}
\title{Reestimate genetic map given a global genotyping error}
\usage{
est_full_hmm_with_global_error(input.map, error = NULL, tol = 0.001,
  restricted = TRUE, th.prob = 0.95, verbose = FALSE)
}
\arguments{
\item{input.map}{an object of class \code{mappoly.map}}

\item{error}{the assumed global error rate (default = NULL)}

\item{tol}{the desired accuracy (default = 10e-04)}

\item{restricted}{if \code{TRUE} (default), restricts the prior to the 
possible classes under mendelian non double-reduced segregation 
given dosage of the parents}

\item{th.prob}{the threshold for using global error or genotype 
probability distribution contained in the data set (default = 0.95)}

\item{verbose}{if \code{TRUE}, current progress is shown; if
\code{FALSE} (default), no output is produced}
}
\value{
An object of class 'mappoly.map' with the following structure:
\item{m}{the ploidy level}
\item{mrk.names}{the names of markers present in the sequence}
\item{data.name}{name of the dataset of class \code{mappoly.data}}
\item{ph.thres}{the LOD threshold used to define the linkage phase configurations to test}
\item{maps}{a list containing the sequence of markers, their recombination fractions,
the linkage phase configuration for all markers in both parents P and Q and the 
map's joint likelihood}
}
\description{
This function considers a global error when reestimating
a genetic map using Hidden Markov models
}
\examples{
  \dontrun{
    seq1.20<-make_seq_mappoly(hexafake, 1:20)
    counts<-cache_counts_twopt(seq1.20,
                               get.from.web=TRUE)
    subset.pairs<-est_pairwise_rf(seq1.20, counts,
                               n.clusters=1)
    subset.map <- est_rf_hmm_sequential(input.seq  = seq1.20,
                                        thres.twopt = 5,
                                        thres.hmm = 10,
                                        extend.tail = 10,
                                        tol = 0.1,
                                        tol.final = 10e-4,
                                        twopt = subset.pairs,
                                        verbose = TRUE)
    subset.map
    plot(subset.map)      
    subset.map.reest<-est_full_hmm_with_global_error(subset.map, 
                                                     error=0.01, 
                                                     tol=10e-4, 
                                                     verbose = TRUE)
    subset.map.reest
    plot(subset.map.reest)
    }

}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400378}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
