% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/reestimate_map_with_prior_genotypic_distribution.R
\name{est_full_hmm_with_prior_dist}
\alias{est_full_hmm_with_prior_dist}
\title{Reestimate genetic map using dosage prior distribution}
\usage{
est_full_hmm_with_prior_dist(input.map, phase.config = "best",
  tol = 0.001, verbose = TRUE)
}
\arguments{
\item{input.map}{an object of class \code{mappoly.map}}

\item{phase.config}{which phase configuration should be used. "best" (default) 
will choose the maximum likelihood configuration}

\item{tol}{the desired accuracy (default = 10e-04)}

\item{verbose}{if \code{TRUE}, current progress is shown; if
\code{FALSE} (default), no output is produced}
}
\value{
An object of class 'mappoly.map' with the following structure:
\item{m}{the ploidy level}
\item{mrk.names}{the names of markers present in the sequence}
\item{data.name}{name of the dataset of class \code{mappoly.data}}
\item{ph.thres}{the LOD threshold used to define the linkage phase configurations to test}
\item{maps}{a list containing the sequence of markers, their recombination fractions,
the linkage phase configuration for all markers in both parents P and Q and the 
map's joint likelihood}
}
\description{
This function considers dosage prior distribution when reestimating
a genetic map using Hidden Markov models
}
\examples{
  \dontrun{
  
  solcap.p<-vector("list", 12)
  names(solcap.p)<-names(solcap.dose.map)
  for(i in 1:12)
    solcap.p[[i]] <- est_full_hmm_with_prior_dist(solcap.dose.map[[i]], verbose = FALSE)
 w<-NULL
 for(i in 1:12)
   w<-c(w, c(solcap.dose.map[i], 
             solcap.p[i]))
             
 names(w) <- apply(expand.grid(c("dose", "prior"), paste0("LG_", 1:12), 
                             stringsAsFactors = FALSE)[,2:1], 1, paste, 
                 collapse = "_")
                 
 op <- par(cex.axis = .7)
 plot_map_list(w, horiz = FALSE, col = rep(gg_color_hue(2), 12))
 par(op)
 legend("bottomright", legend = c("Dosage based", "Prior"), pch=15, col = rep(gg_color_hue(2)))
  
}

}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400378}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
