% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_map_utils.R
\name{est_haplo_hmm}
\alias{est_haplo_hmm}
\title{Estimate a genetic map given a sequence of block markers}
\usage{
est_haplo_hmm(
  m,
  n.mrk,
  n.ind,
  haplo,
  emit = NULL,
  rf_vec,
  verbose,
  use_H0 = FALSE,
  tol
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Estimate a genetic map given a sequence of block markers
}
\keyword{internal}
