% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_rf.R
\name{est_pairwise_rf}
\alias{est_pairwise_rf}
\alias{print.poly.est.two.pts.pairwise}
\alias{plot.poly.est.two.pts.pairwise}
\title{Pairwise two-point analysis}
\usage{
est_pairwise_rf(input.seq, count.cache, n.clusters = 1,
  tol = .Machine$double.eps^0.25, mrk.pairs = NULL,
  batch.size = NULL, verbose = TRUE)

\method{print}{poly.est.two.pts.pairwise}(x, ...)

\method{plot}{poly.est.two.pts.pairwise}(x, first.mrk, second.mrk, ...)
}
\arguments{
\item{input.seq}{an object of class \code{mappoly.sequence}}

\item{count.cache}{an object of class \code{cache.info} containing
pre-computed genotype frequencies, obtained with
\code{\link[mappoly]{cache_counts_twopt}}}

\item{n.clusters}{Number of parallel processes (cores) to spawn (default = 1)}

\item{tol}{the desired accuracy. See \code{optimize()} for details}

\item{mrk.pairs}{a matrix of dimensions 2*N, containing N
pairs of markers to be analyzed. If \code{NULL} (default), all pairs are
considered}

\item{batch.size}{The number of pairs that shold be analyzed in parallel.
If \code{NULL} (default), all pairs are analyized}

\item{verbose}{If \code{TRUE} (default), current progress is shown; if
\code{FALSE}, no output is produced}
}
\value{
An object of class \code{poly.est.two.pts.pairwise} which
    is a list  containing the following components:
    \item{data.name}{name of the object of class
        \code{mappoly.data} with the raw data}
    \item{n.mrk}{number of markers in the sequence}
    \item{seq.num}{a \code{vector} containing the (ordered) indices
        of markers inthe sequence, according to the input file}
    \item{pairwise}{a list of size
    \code{choose(length(input.seq$seq.num), 2)}, each of them containing: a matrix
    with the linkage phase configuration i.e. the number of
    homologous that share alelels; the LOD Score relative to each
    one of these configurations; the recombination fraction
    estimated for each configuration; and their associated LOD
    score}
    \code{chisq.pval.thres}{threshold used to perform the segregation tests}
    \code{chisq.pval}{p-values associated with the performed segregation tests}
}
\description{
Performs the two-point pairwise analysis, accounting for the recombination fractions
between all pairwise markers given one or more linkage phase configurations.
}
\examples{
  \dontrun{
    data(hexafake)
    all.mrk<-make_seq_mappoly(hexafake, 'all')
    red.mrk<-elim_redundant(all.mrk)
    unique.mrks<-make_seq_mappoly(red.mrk)
    counts.web<-cache_counts_twopt(unique.mrks, get.from.web = TRUE)
    all.pairs<-est_pairwise_rf(input.seq = unique.mrks,
                               count.cache = counts.web,
                               n.clusters = 16,
                               verbose=TRUE)
   all.pairs
   plot(all.pairs, 90, 91)
   }
   
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400378}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
