% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_hmm.R
\name{est_rf_hmm}
\alias{est_rf_hmm}
\alias{print.mappoly.map}
\alias{plot.mappoly.map}
\title{Multipoint analysis using Hidden Markov Models in autopolyploids}
\usage{
est_rf_hmm(input.seq, input.ph = NULL, thres = 0.5, twopt = NULL,
  verbose = FALSE, tol = 1e-04, est.given.0.rf = FALSE,
  reestimate.single.ph.configuration = TRUE, high.prec = TRUE)

\method{print}{mappoly.map}(x, detailed = FALSE, ...)

\method{plot}{mappoly.map}(x, phase = TRUE, col.P = "#e41a1c",
  col.Q = "#377eb8", mrk.names = FALSE, cex = 1, config = "best",
  ...)
}
\arguments{
\item{input.seq}{an object of class \code{mappoly.sequence}}

\item{input.ph}{an object of class \code{two.pts.linkage.phases}. 
If not available (default = NULL), it will be computed}

\item{thres}{the LOD threshold used to determine if the linkage phases
compared via two-point analysis should be considered. Smaller 
values will result in smaller number of linkage phase 
configurations}

\item{twopt}{an object of class \code{poly.est.two.pts.pairwise}
containing two-point information}

\item{verbose}{if \code{TRUE}, current progress is shown; if
\code{FALSE} (default), no output is produced}

\item{tol}{the desired accuracy (default = 1e-04)}

\item{est.given.0.rf}{logical. If TRUE returns a map forcing all
recombination fractions equals to 0 (1e-5, for internal use only. Default = FALSE)}

\item{reestimate.single.ph.configuration}{logical. If \code{TRUE}
returns a map without reestimating the map parameters for cases
where there is only one possible linkage phase configuration. 
This argument is intended to be used in a sequential map contruction}

\item{high.prec}{logical. If \code{TRUE} (default) uses high precision 
long double numbers in the HMM procedure}

\item{x}{an object of the class \code{mappoly.map}}

\item{detailed}{logical. if TRUE, prints the linkage phase configuration and the marker 
position for all maps. if FALSE (default), prints a map summary}

\item{...}{currently ignored}

\item{phase}{logical. If \code{TRUE} (default) plots the phase configuration
for both parents}

\item{mrk.names}{if TRUE, marker names are displayed (default = FALSE)}

\item{cex}{The magnification to be used for marker names}

\item{config}{should be \code{'best'} or the position of the
configuration to be plotted. If \code{'best'}, plot the configuration
with the highest likelihood}

\item{col.cte.P}{a single character string or a vector of strings with size 
equal to the number of markers in the map indicating the color of the allelic
 variants for parent P (default = 'red')}

\item{col.cte.Q}{a single character string or a vector of strings with size 
equal to the number of markers in the map indicating the color of the allelic
 variants for parent Q (default = 'blue')}
}
\value{
An object of class \code{mappoly.map} with the following structure:
\item{m}{the ploidy level}
\item{mrk.names}{the names of markers present in the sequence}
\item{data.name}{name of the dataset of class \code{mappoly.data}}
\item{ph.thres}{the LOD threshold used to define the linkage phase configurations to test}
\item{maps}{a list containing the sequence of markers, their recombination fractions,
the linkage phase configuration for all markers in both parents P and Q and the 
map's joint likelihood}
}
\description{
Performs the multipoint analysis proposed by \cite{Mollinari and
Garcia (2019)} in a sequence of markers
}
\details{
This function first enumerates a set of linkage phase configurations
 based on two-point recombination fraction information using a threshold
 provided by the user (argument \code{thresh}). After that, for each
 configuration, it reconstructs the genetic map using the
 HMM approach described in Mollinari and Garcia (2019). As result, it returns 
 the multipoint likelihood for each configuration in form of LOD Score comparing 
 each configuration to the most likely one. It is recommended to use a small number 
 of markers (e.g. 50 markers for hexaploids) since the possible linkage 
 phase combinations bounded only by the two-point information can be huge. 
 Also, it can be quite sensible to small changes in \code{'thresh'}. 
 For higher number of markers, please see \code{\link[mappoly]{est_rf_hmm_sequential}}.
}
\examples{
 \dontrun{
    data(hexafake)
    mrk.subset<-make_seq_mappoly(hexafake, 1:50)
    red.mrk<-elim_redundant(mrk.subset)
    unique.mrks<-make_seq_mappoly(red.mrk)
    counts.web<-cache_counts_twopt(unique.mrks, get.from.web = TRUE)
    subset.pairs<-est_pairwise_rf(input.seq = unique.mrks,
                                  count.cache = counts.web,
                                  n.clusters = 1,
                                  verbose=TRUE)

    ## Estimating subset map with a low tolerance for the E.M. procedure
    subset.map <- est_rf_hmm(input.seq = unique.mrks,
                             thres = 2,
                             twopt = subset.pairs,
                             verbose = TRUE,
                             tol = 0.1,
                             est.given.0.rf = FALSE)

    ## Re-estimating the map with the most likely configuration
    subset.map1 <- est_rf_hmm_single(input.seq = unique.mrks,
                                    input.ph.single = subset.map$maps[[1]]$seq.ph,
                                    tol = 10e-3,
                                    verbose = TRUE)

    subset.map$maps[[1]]$seq.ph <- subset.map1$seq.ph
    
    plot(subset.map)

     ## Retrieving simulated linkage phase
     ph.P <- maps.hexafake[[1]]$maps[[1]]$seq.ph$P
     ph.Q <- maps.hexafake[[1]]$maps[[1]]$seq.ph$Q

     ## Estimated linkage phase
     ph.P.est <- subset.map$maps[[1]]$seq.ph$P
     ph.Q.est <- subset.map$maps[[1]]$seq.ph$Q

     compare_haplotypes(m = 6, h1 = ph.P[names(ph.P.est)], h2 = ph.P.est)
     compare_haplotypes(m = 6, h1 = ph.Q[names(ph.Q.est)], h2 = ph.Q.est)
   }

}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    https://doi.org/10.1534/g3.119.400378
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{internal}
