% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_map_hmm.R
\name{filter_map_at_hmm_thres}
\alias{filter_map_at_hmm_thres}
\title{remove maps under a certain threshold}
\usage{
filter_map_at_hmm_thres(map, thres.hmm)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
remove maps under a certain threshold
}
\keyword{internal}
