% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_missing}
\alias{filter_missing}
\title{Filter missing genotypes}
\usage{
filter_missing(input.data, filter.thres = 0.8, inter = TRUE)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Filter missing genotypes
}
\keyword{internal}
