% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotype_map_utils.R
\name{generate_all_link_phases_elim_equivalent_haplo}
\alias{generate_all_link_phases_elim_equivalent_haplo}
\title{Generates all possible linkage phases between two blocks of markers,
eliminating equivalent configurations, i.e., configurations with the
same likelihood. If the argument \code{reduce} is TRUE (default),
the \code{thresh.LOD.ph} value defined in the \code{rf_list_to_matrix}
function is used. Then, only markers that presented phases
with LOD differences higher than \code{thresh.LOD.ph} will be used,
in order to reduce the number of phases to test.}
\usage{
generate_all_link_phases_elim_equivalent_haplo(block1, block2, rf.matrix,
  m, reduce = TRUE, max.inc = NULL)
}
\arguments{
\item{block1}{submap with markers of the first block}

\item{block2}{submap with markers of the second block, or just a single marker identified by its position in the \code{mappoly.data} object}

\item{rf.matrix}{matrix obtained with the function \code{rf_list_to_matrix}
using the parameter \code{shared.alleles = TRUE}}

\item{m}{ploidy level (i.e. 4, 6 and so on)}

\item{max.inc}{maximum number of allowed inconsistencies}
}
\description{
Generates all possible linkage phases between two blocks of markers,
eliminating equivalent configurations, i.e., configurations with the
same likelihood. If the argument \code{reduce} is TRUE (default),
the \code{thresh.LOD.ph} value defined in the \code{rf_list_to_matrix}
function is used. Then, only markers that presented phases
with LOD differences higher than \code{thresh.LOD.ph} will be used,
in order to reduce the number of phases to test.
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu} and Gabriel Gesteira, \email{gabrielgesteira@usp.br}
}
\keyword{internal}
