% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reestimate_map_with_global_genotypic_error.R
\name{genotyping_global_error}
\alias{genotyping_global_error}
\title{Introduces genotyping error into a full informative vector}
\usage{
genotyping_global_error(x, error = 0.01, th.prob = 0.999)
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Introduces genotyping error into a full informative vector
}
\keyword{internal}
