% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dosage_type}
\alias{get_dosage_type}
\title{Get dosage type in a sequence}
\usage{
get_dosage_type(input.seq)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Get dosage type in a sequence
}
\keyword{internal}
