% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_genomic_order}
\alias{get_genomic_order}
\title{Get genomic order of the markers}
\usage{
get_genomic_order(input.seq)
}
\arguments{
\item{input.seq}{a sequence object of class \code{mappoly.sequence}}
}
\description{
This functions gets the genomic order of the markers, if provided.
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
\keyword{genomic}
