% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_genomic_order}
\alias{get_genomic_order}
\title{Get the genomic position of markers in a sequence}
\usage{
get_genomic_order(input.seq, verbose = TRUE)
}
\arguments{
\item{input.seq}{a sequence object of class \code{mappoly.sequence}}

\item{verbose}{if \code{TRUE} (default), the current progress is shown; if
\code{FALSE}, no output is produced}
}
\description{
This functions gets the genomic position of markers in a sequence and
return an ordered data frame with the name and position of each marker
}
\examples{
 s1<-make_seq_mappoly(tetra.solcap, "all")
 o1<-get_genomic_order(s1)
 head(o1)
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
