% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_memory}
\alias{get_memory}
\title{Get used memory (unix only)}
\usage{
get_memory()
}
\arguments{
\item{void}{interfunction to be documented}
}
\description{
Get used memory (unix only)
}
\keyword{internal}
