% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_from_polymapR.R
\name{import_phased_maplist_from_polymapR}
\alias{import_phased_maplist_from_polymapR}
\title{Import phased map list from polymapR}
\usage{
import_phased_maplist_from_polymapR(maplist, mappoly.data, ploidy = NULL)
}
\arguments{
\item{maplist}{a list of phased maps obtained using function 
\code{create_phased_maplist} from package \code{polymapR}}

\item{mappoly.data}{a dataset used to obtain \code{maplist}, 
converted into class \code{mappoly.data}}

\item{ploidy}{the ploidy level}
}
\description{
Function to import phased map lists from polymapR
}
\details{
See examples at \url{https://rpubs.com/mmollin/tetra_mappoly_vignette}.
}
\references{
Bourke PM et al: (2019) PolymapR — linkage analysis and genetic map 
    construction from F1 populations of outcrossing polyploids. 
    _Bioinformatics_ 34:3496–3502.
    \doi{10.1093/bioinformatics/bty1002}

    Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \doi{10.1534/g3.119.400378}
}
\author{
Marcelo Mollinari \email{mmollin@ncsu.edu}
}
