% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.prob.data}
\alias{is.prob.data}
\title{Is it a probability dataset?}
\usage{
is.prob.data(x)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Is it a probability dataset?
}
\keyword{internal}
