% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pairs.R
\name{make_pairs_mappoly}
\alias{make_pairs_mappoly}
\title{Subset pairwise recombination fractions}
\usage{
make_pairs_mappoly(input.twopt, input.seq)
}
\arguments{
\item{input.twopt}{an object of class \code{poly.est.two.pts.pairwise}.}

\item{input.seq}{an object of class \code{mappoly.sequence}.
It must be contained in 'input.twopt'}
}
\value{
a subset of \code{'input.twopt'}. This object is also
    of class \code{poly.est.two.pts.pairwise}.
}
\description{
Get a subset of an object of class \code{'poly.est.two.pts.pairwise'}, i.e.,
 recombination fraction and LOD score statistics for all possible linkage
 phase combinations based in a sequence of markers.
}
\examples{
 \dontrun{
    data(hexafake)
    all.mrk<-make_seq_mappoly(hexafake, 'all')
    red.mrk<-elim_redundant(all.mrk)
    unique.mrks<-make_seq_mappoly(red.mrk)
    counts.web<-cache_counts_twopt(unique.mrks, get.from.web = TRUE)
    all.pairs<-est_pairwise_rf(input.seq = unique.mrks,
                               count.cache = counts.web,
                               n.clusters = 16,
                               verbose=TRUE)

    ## Full recombination fraction matrix
    mat.full<-rf_list_to_matrix(input.twopt = all.pairs)
    plot(mat.full)

    lgs <- group_mappoly(input.mat = mat.full,
                         input.seq = unique.mrks,
                         expected.groups = 3,
                         verbose=TRUE)
    lgs
    plot(lgs)
    lg1 <- make_seq_mappoly(lgs, 1)
    lg2 <- make_seq_mappoly(lgs, 2)
    lg3 <- make_seq_mappoly(lgs, 3)

    ##Plot matrices
    p1<-make_pairs_mappoly(input.seq = lg1, input.twopt = all.pairs)
    p2<-make_pairs_mappoly(input.seq = lg2, input.twopt = all.pairs)
    p3<-make_pairs_mappoly(input.seq = lg3, input.twopt = all.pairs)

    m1<-rf_list_to_matrix(input.twopt = p1)
    m2<-rf_list_to_matrix(input.twopt = p2)
    m3<-rf_list_to_matrix(input.twopt = p3)

    op<-par(mfrow = c(1,3), pty = "s")
    plot(m1, main.text = "LG1")
    plot(m2, main.text = "LG2")
    plot(m3, main.text = "LG3")
    par(op)
   }
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2018) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_. \url{https://doi.org/10.1101/415232}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
