% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_maps.R
\name{merge_maps}
\alias{merge_maps}
\title{Merge two maps}
\usage{
merge_maps(input.map1, input.map2, twopt, thres.twopt = 3,
  thres.hmm = 10, genoprob.map1 = NULL, phase.config.map1 = "best",
  phase.config.map2 = "best", tol = 0.001)
}
\arguments{
\item{input.map1}{an object of class \code{mappoly.map}, which comprehends the first map to be merged}

\item{input.map2}{an object of class \code{mappoly.map}, which comprehends the second map to be merged}

\item{twopt}{an object of class \code{poly.est.two.pts.pairwise}
containing the two-point information}

\item{thres.twopt}{the threshold used to determine if the linkage
phases compared via two-point analysis should be considered 
for the search space reduction (default = 3)}

\item{thres.hmm}{the threshold used to determine which linkage 
phase configurations should be retuned (default = 10)}

\item{genoprob.map1}{an object of class \code{mappoly.genoprob} containing the 
genotype probabilities for the first map}

\item{phase.config.map1}{which phase configuration should be used for \code{input.map1}. "best" (default) 
will choose the configuration associated with the maximum likelihood}

\item{phase.config.map2}{which phase configuration should be used for \code{input.map2}. "best" (default) 
will choose the configuration associated with the maximum likelihood}

\item{tol}{the desired accuracy (default = 10e-04)}
}
\value{
An object of class 'mappoly.map' with the following structure:
\item{m}{the ploidy level}
\item{mrk.names}{the names of markers present in the sequence}
\item{data.name}{name of the dataset of class \code{mappoly.data}}
\item{ph.thres}{the LOD threshold used to define the linkage phase configurations to test}
\item{maps}{a list containing the sequence of markers, their recombination fractions,
the linkage phase configuration for all markers in both parents P and Q and the 
map's joint likelihood}
}
\description{
Estimates the linkage phase and recombination fraction between two pre-built maps 
and creates a new map by merging them.
}
\details{
\code{merge_maps} uses two-point information, under a given LOD threshold, to reduce the 
linkage phase search space. The remaining linkage phases are tested using the genotype 
probabilities
}
\examples{
\dontrun{
#### Tetraploid example #####
map1<-get_submap(solcap.dose.map[[1]], 1:5)
map2<-get_submap(solcap.dose.map[[1]], 6:15)
full.map<-get_submap(solcap.dose.map[[1]], 1:15)
s<-make_seq_mappoly(tetra.solcap, full.map$maps[[1]]$seq.num)
counts<-cache_counts_twopt(input.seq = s, get.from.web = TRUE)
twopt <- est_pairwise_rf(input.seq = s, count.cache = counts)
genoprob.map1 <- calc_genoprob(map1)
merged.maps<-merge_maps(input.map1 = map1, 
                        input.map2 = map2, 
                        twopt = twopt,
                        thres.twopt = 3,
                        genoprob.map1 = genoprob.map1)

plot(merged.maps, mrk.names = TRUE)                       
plot(full.map, mrk.names = TRUE)                       
best.phase <- merged.maps$maps[[1]]$seq.ph
names.id<-names(best.phase$P)
plot_compare_haplotypes(m = 6,
                        hom.allele.p1 = best.phase$P[names.id],
                        hom.allele.q1 = best.phase$Q[names.id],
                        hom.allele.p2 = full.map$maps[[1]]$seq.ph$P[names.id],
                        hom.allele.q2 = full.map$maps[[1]]$seq.ph$Q[names.id])  
                        
#### Hexaploid example #####
map1<-get_submap(maps.hexafake[[1]], 1:5)
map2<-get_submap(maps.hexafake[[1]], 6:15)
full.map<-get_submap(maps.hexafake[[1]], 1:15)
s<-make_seq_mappoly(hexafake, full.map$maps[[1]]$seq.num)
counts<-cache_counts_twopt(input.seq = s, get.from.web = TRUE)
twopt <- est_pairwise_rf(input.seq = s, count.cache = counts)
genoprob.map1 <- calc_genoprob(map1)
merged.maps<-merge_maps(input.map1 = map1, 
                        input.map2 = map2, 
                        twopt = twopt,
                        thres.twopt = 3,
                        genoprob.map1 = genoprob.map1)

plot(merged.maps, mrk.names = TRUE)                       
plot(full.map, mrk.names = TRUE)                       
best.phase <- merged.maps$maps[[1]]$seq.ph
names.id<-names(best.phase$P)
plot_compare_haplotypes(m = 6,
                        hom.allele.p1 = best.phase$P[names.id],
                        hom.allele.q1 = best.phase$Q[names.id],
                        hom.allele.p2 = full.map$maps[[1]]$seq.ph$P[names.id],
                        hom.allele.q2 = full.map$maps[[1]]$seq.ph$Q[names.id])  
}

}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
