% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_compare_haplotypes}
\alias{plot_compare_haplotypes}
\title{Plot two overlapped haplotypes}
\usage{
plot_compare_haplotypes(
  m,
  hom.allele.p1,
  hom.allele.q1,
  hom.allele.p2 = NULL,
  hom.allele.q2 = NULL
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Plot two overlapped haplotypes
}
\keyword{internal}
