% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_list.R
\name{plot_map_list}
\alias{plot_map_list}
\title{Plot a genetic map}
\usage{
plot_map_list(map.list, horiz = TRUE, col = "lightgray",
  title = "Linkage Groups")
}
\arguments{
\item{map.list}{A list of objects or a single object of class \code{mappoly.map}}

\item{horiz}{logical. If FALSE, the maps are drawn vertically with the first map to the left. 
If TRUE  (default), the maps are drawn horizontally with the first at the bottom}

\item{col}{a vector of colors for the bars or bar components (default = 'lightgrey')
\code{ggstyle} produces maps using the default \code{ggplot} color palette}

\item{title}{a title (string) for the maps (default = 'Linkage Groups')}
}
\value{
A \code{data.frame} object containing the name of the markers and their genetic position
}
\description{
This function draws genetic linkage map(s) generated by some of the \code{MAPpoly} functions.
The map(s) should be passed as a single object or a list of objects of class \code{mappoly.map}.
}
\examples{
 \dontrun{
 ## hexafake map
 plot_map_list(maps.hexafake, horiz = FALSE)
 plot_map_list(maps.hexafake, col = c("#999999", "#E69F00", "#56B4E9"))
 
 ## solcap map
 plot_map_list(solcap.map, col = "ggstyle")
 
 ## Comparing mapping approaches
 
 w<-NULL
 for(i in 1:12)
   w<-c(w, c(solcap.dose.map[i], 
             solcap.prior.map[i],
             solcap.err.map[i]))
             
 names(w) <- apply(expand.grid(c("dose", "prior", "error"), paste0("LG_", 1:12), 
                             stringsAsFactors = FALSE)[,2:1], 1, paste, 
                 collapse = "_")
                 
 op <- par(cex.axis = .7)
 plot_map_list(w, horiz = FALSE, col = rep(gg_color_hue(3), 12))
 par(op)
 legend("bottomright", legend = c("Dosage based", "Prior", "Error"), 
         pch=15, col = rep(gg_color_hue(3)))
 }
 
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400378}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
