% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_info.R
\name{plot_mrk_info}
\alias{plot_mrk_info}
\title{Plot marker information}
\usage{
plot_mrk_info(input.data, mrk)
}
\arguments{
\item{input.data}{an object of class \code{mappoly.data}}

\item{mrk}{marker name or position in the dataset}
}
\description{
Plots summary statistics for a given marker.
}
\examples{
plot_mrk_info(tetra.solcap.geno.dist, 2680)
plot_mrk_info(tetra.solcap.geno.dist, "solcap_snp_c2_23828")

}
