% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_info.R
\name{plot_mrk_info}
\alias{plot_mrk_info}
\title{Plot marker information}
\usage{
plot_mrk_info(input.data, mrk)
}
\arguments{
\item{input.data}{an object of class \code{mappoly.data}}

\item{mrk}{marker name or position in the dataset}
}
\description{
Plots summary statistics for a given marker
}
\examples{
 plot_mrk_info(tetra.solcap.geno.dist, 2680)
 plot_mrk_info(tetra.solcap.geno.dist, "solcap_snp_c2_23828")
    
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400378}
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
