% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_progeny_dosage_change.R
\name{plot_progeny_dosage_change}
\alias{plot_progeny_dosage_change}
\title{Look at genotypes that were imputed or changed by the HMM chain given a level of global genotypic error}
\usage{
plot_progeny_dosage_change(map_list, error, verbose = T)
}
\arguments{
\item{map_list}{a list of multiple \code{mappoly.map.list}}

\item{error}{error rate used in global error in the `calc_genoprob_error()`}

\item{verbose}{T or F for `calc_genoprob_error()` and `calc_homologprob()`}
}
\value{
A ggplot of the changed and imputed genotypic dosages
}
\description{
Outputs a graphical representation ggplot with the percent of data changed
}
\examples{
    plot_progeny_dosage_change(map_list=MAPs, error=0.05)

}
\author{
Jeekin Lau, \email{jzl0026@tamu.edu}, with optimization by Cristiane Taniguti, \email{chtaniguti@tamu.edu}
}
