% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mappoly_vcf.R
\name{read_vcf}
\alias{read_vcf}
\title{Data Input VCF}
\usage{
read_vcf(
  file.in,
  filter.non.conforming = TRUE,
  parent.1,
  parent.2,
  ploidy = NA,
  thresh.line = 0.05,
  min.gt.depth = 0,
  min.av.depth = 0,
  max.missing = 1,
  elim.redundant = TRUE
)
}
\arguments{
\item{file.in}{a character string with the name of (or full path to) the input file which contains the data (VCF format)}

\item{filter.non.conforming}{if \code{TRUE} (default) exclude samples with non 
expected genotypes under random chromosome pairing and no double reduction}

\item{parent.1}{a character string containing the name of parent 1}

\item{parent.2}{a character string containing the name of parent 2}

\item{ploidy}{the species ploidy (optional, it will be automatically detected)}

\item{thresh.line}{threshold used for p-values on segregation test (default = 0.05)}

\item{min.gt.depth}{minimum genotype depth to keep information. If the genotype depth is below \code{min.gt.depth},
it will be replaced with NA (default = 0)}

\item{min.av.depth}{minimum average depth to keep markers (default = 0)}

\item{max.missing}{maximum proportion of missing data to keep markers (range = 0-1; default = 1)}

\item{elim.redundant}{logical. If \code{TRUE} (default), removes redundant markers
during map construction, keeping them annotated to export to the final map.}
}
\value{
An object of class \code{mappoly.data} which contains a
    list with the following components:
    \item{m}{ploidy level}
    \item{n.ind}{number individuals}
    \item{n.mrk}{total number of markers}
    \item{ind.names}{the names of the individuals}
    \item{mrk.names}{the names of the markers}
    \item{dosage.p}{a vector containing the dosage in
      parent P for all \code{n.mrk} markers}
    \item{dosage.q}{a vector containing the dosage in
      parent Q for all \code{n.mrk} markers}
    \item{sequence}{a vector indicating which sequence each marker
      belongs. Zero indicates that the marker was not assigned to any
      sequence}
    \item{sequence.pos}{Physical position of the markers into the
      sequence}
    \item{seq.ref}{Reference base used for each marker (i.e. A, T, C, G)}
    \item{seq.alt}{Alternative base used for each marker (i.e. A, T, C, G)}
    \item{prob.thres}{(unused field)}
    \item{geno.dose}{a matrix containing the dosage for each markers (rows) 
      for each individual (columns). Missing data are represented by 
      \code{ploidy_level + 1}}
    \item{nphen}{(unused field)}
    \item{phen}{(unused field)}
    \item{all.mrk.depth}{DP information for all markers on VCF file}
    \item{chisq.pval}{a vector containing p-values related to the chi-squared 
    test of mendelian segregation performed for all markers}
    \item{unique.seq}{if elim.redundant=TRUE, holds the object of class 'mappoly.unique.seq'}
    \item{kept}{if elim.redundant=TRUE, holds all non-redundant markers}
    \item{elim.correspondence}{if elim.redundant=TRUE, holds all non-redundant markers and
its equivalence to the redundant ones}
}
\description{
Reads an external VCF file. This function accepts version 4.0 and higher, and creates an object of class \code{mappoly.data}
}
\details{
This function can handle .vcf files of version 4.0 and higher. The ploidy can be automatically detected, but it is highly
recommended that you inform it to check for mismatches. All individual and marker names will be kept as they are in the .vcf file.
}
\examples{
\dontrun{
download.file("https://github.com/mmollina/MAPpoly_vignettes/raw/master/data/sweetpotato_aligned_trifida_chr1.vcf.gz", destfile = 'chr1.vcf.gz')
dat.dose.vcf = read_vcf(file = 'chr1.vcf.gz', parent.1 = "PARENT1", parent.2 = "PARENT2")
}
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2019) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _G3: Genes, Genomes, Genetics_. 
    \url{https://doi.org/10.1534/g3.119.400378}
}
\author{
Gabriel Gesteira, \email{gabrielgesteira@usp.br}
}
