% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rev_map}
\alias{rev_map}
\title{Reverse map}
\usage{
rev_map(input.map)
}
\arguments{
\item{input.map}{An object of class \code{mappoly.map}}
}
\description{
Provides a reverse version of a given map.
}
\references{
Mollinari, M., and Garcia, A.  A. F. (2017) Linkage
    analysis and haplotype phasing in experimental autopolyploid
    populations with high ploidy level using hidden Markov
    models, _submited_
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
