% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rev_map}
\alias{rev_map}
\title{Reverse map}
\usage{
rev_map(input.map)
}
\arguments{
\item{input.map}{an object of class \code{mappoly.map}}
}
\description{
Provides the reverse version of a given map.
}
\author{
Marcelo Mollinari, \email{mmollin@ncsu.edu}
}
