% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_utils.R
\name{sim_cross_two_informative_parents}
\alias{sim_cross_two_informative_parents}
\title{Simulate mapping population (tow parents)}
\usage{
sim_cross_two_informative_parents(
  m,
  n.mrk,
  rf.vec,
  n.ind,
  hom.allele.p,
  hom.allele.q,
  prob.P = NULL,
  prob.Q = NULL,
  seed = NULL
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Simulate mapping population (tow parents)
}
\keyword{internal}
