% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_mappoly}
\alias{split_mappoly}
\title{Split map into sub maps given a gap threshold}
\usage{
split_mappoly(
  input.map,
  gap.threshold = 5,
  size.rem.cluster = 1,
  phase.config = "best",
  tol.final = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{void}{internal function to be documented}
}
\description{
Split map into sub maps given a gap threshold
}
\keyword{internal}
