% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_maps}
\alias{summary_maps}
\title{Summary map}
\usage{
summary_maps(map.object)
}
\arguments{
\item{map.object}{a list of objects of class \code{mappoly.map}}
}
\value{
a dataframe containing a brief summary of all maps contained in \code{map.object}
}
\description{
This function generates a brief summary table of a list of \code{mappoly.map} objects
}
\examples{
 \dontrun{
(tetra.sum <- summary_maps(solcap.err.map))
formattable::formattable(tetra.sum)
(hexa.sum <- summary_maps(maps.hexafake))
formattable::formattable(hexa.sum)
}

}
\author{
Gabriel Gesteira, \email{gabrielgesteira@usp.br}
}
