% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summary_maps}
\alias{summary_maps}
\title{Summary maps}
\usage{
summary_maps(map.list, verbose = TRUE)
}
\arguments{
\item{map.list}{a list of objects of class \code{mappoly.map}}

\item{verbose}{if \code{TRUE} (default), the current progress is shown; if
\code{FALSE}, no output is produced}
}
\value{
a data frame containing a brief summary of all maps contained in \code{map.list}
}
\description{
This function generates a brief summary table of a list of \code{mappoly.map} objects
}
\examples{
tetra.sum <- summary_maps(solcap.err.map)
tetra.sum
}
\author{
Gabriel Gesteira, \email{gabrielgesteira@usp.br}
}
